### 2019-07-23 #1590 @whedon
[PRE REVIEW]: Virtual Scanner:  MRI on a Browser
**Submitting author:** @imr-framework (<a href="http://orcid.org/0000-0002-3776-4114">Sairam Geethanath</a>)
**Repository:** <a href="https://github.com/imr-framework/virtual-scanner" target ="_blank">https://github.com/imr-framework/virtual-scanner</a>
**Version:** 1.0.0
