### 2019-07-23 #1590 @kyleniemeyer
Hi @sairamgeethanath, thanks for your submission to JOSS! I'm about to assign an editor to handle your submission, but first:

JOSS only considers software with a clear research application. Your submission looks interesting, but the paper and documentation seem to focus more on educational/training applications. We think your tool does in fact have research applications, but they aren't currently clear. Can you revise your paper to discuss this more?
