### 2019-07-24 efd6d1a
Fix SE results not showing up
