### 2019-07-24 #1590 @arokem
👋 @thechymera : would you be willing to review this paper for JOSS?
