### 2019-07-26 #1590 @arokem
@TheChymera : would you be willing to review this submission for JOSS?
