### 2019-07-26 12b6a47
JOSS submission - edit 1
