+++ paper/paper.md
@@ -2,7 +2,7 @@
 title: 'Virtual Scanner: MRI on a Browser'
 
 tags:
-  - Magnetic Resonance Imaging
+  - Accessible Magnetic Resonance Imaging
   - Web Application
   - Python
   - Pulseq
@@ -25,7 +25,7 @@ authors:
     affiliation: 1
   - name: Nishika Girish
     affiliation: 2
-  - name: John Thomas Vaughan
+  - name: John Thomas Vaughan Jr.
     orcid: 0000-0002-6933-3757
     affiliation: 1
 
@@ -35,48 +35,48 @@ affiliations:
  - name: Plum Grove Junior High, Palatine, IL, USA
    index: 2
  
-date: 22 July 2019
+date: 26 July 2019
 
 bibliography: paper.bib
 ---
 # Summary
-Magnetic Resonance Imaging (MRI) is a medical imaging modality that provides excellent soft-tissue contrast with high diagnostic value. Well-trained personnel are required for sustaining deployment of MRI systems because of their inherent complexities. In underserved or remote areas, local expertise is insufficiently established, which formed the motivation for this design challenge [@obungoloch_2019]. 
+Magnetic Resonance Imaging (MRI) is a medical imaging modality that provides excellent soft-tissue contrast with high diagnostic value. 
+System simulators are key to method development in MRI because of its inherent complexity. However, most existing MR system-wide simulators are proprietary to manufacturing companies, and an open-source, integrated webtool would help MR researchers around the world develop and share new methodology. Specifically, it would provide vendor-neutral tools for rapidly prototyping MR acquisition methods and hardware components, services for sharing computational methods and resources, and education for under-served areas to develop their own MR expertise [@obungoloch_2019]. Being vendor-neutral would also enable multi-site repeatability studies. To fulfill this need, we designed ``Virtual Scanner`` with three main characteristics in mind:
 
-In response, we built ``Virtual Scanner``[@vsrepo], an MR system simulator employable for both educational and research purposes. On one hand, a free and comprehensive MR education tool can help maximize the long-term influence of hardware in underserved areas. On the other hand, most existing MR system simulators are proprietary to manufacturing companies and an open-source, integrated webtool will help MR researchers around the world develop and share new methodology.
 
-``Virtual Scanner``[@vsrepo] has three main characteristics:
-
-* Modular: modern MRI systems are equipped with multiple coils for producing magnetic fields, radiofrequency (RF) electronics, console software, and related accessories. ``Virtual Scanner`` consists of modules corresponding to steps in the signal chain [@webb2016magnetic]. Trainees can learn about the imaging experiment by choosing modules that cater to their practical needs.
+* Modular: ``Virtual Scanner`` consists of modules corresponding to steps in the MR signal chain [@webb2016magnetic]. Since modern MRI systems contain multiple levels of hardware and software, it is important for researchers to select the appropriate steps for simulation when developing acquisition paradigms, analysis methods, or hardware components. Students may also choose to focus on individual aspects of the experiment.
 
 * Zero-footprint: ``Virtual Scanner`` was implemented as a web application using the FLASK framework [@flask] with backend code in Python. The local server can be set up by running a script or using the command line. Once the server is up, one can access it on a browser without additional steps for installing any software packages. 
 
-* Open-source: existing tools in the MR community are incorporated for effective creation and sharing of resources. Pulseq, a multi-vendor MR pulse sequence format [@layton2017pulseq; @ravi2018pulseq], can be directly fed into the simulator. This is especially helpful when users plan to translate from simulation to experiments in a real scanner, since they may use the same sequence file for both purposes.
[…]
