### 2019-07-27 #1590 @TheChymera
@arokem thank you for your consideration.

As a developer, I am not at all familiar with web applications, and as a user I have always found them unwieldy to integrate into other workflows. It is entirely possible that given this design choice by the authors, my review will be both slower and more critical, than if it were coming from someone who can actually compare the application to what can be expected of web apps in general.

I would encourage you to ask other potential reviewers, including @damaggu . If none other can be found, I could try looking into this no earlier than late August.
