### 2019-07-27 #1590 @arokem
👋 @pgadige: would you be willing to review this paper for JOSS? Your expertise in web applications should come in handy.

👋 @vsoch : would you be willing to review this paper for JOSS? Your unique combination of knowledge about web technologies and MRI makes you uniquely qualified.
