### 2019-07-27 #1590 @vsoch
I can't say no after that! :)

I'd be happy to, I'm also about to review https://github.com/openjournals/joss-reviews/issues/1595 so please keep that in mind. I'm not sure  that @imr-framework would want me to, however, because I don't see any reproducible containers or tests and these would be required (on top of the normal boxes we check off) to give a thumbs up for publication.
