### 2019-07-31 #1590 @arokem
👋 @spinicist : would you be willing to review this paper for JOSS? Your expertise in MRI will be useful here.
