### 2019-08-01 #1590 @spinicist
Hello, what's the time frame on this? I will be on leave soon and am unlikely to get to it before I go, so there could be a substantial delay in my input. If that's acceptable I'm very happy to review.
