### 2019-08-01 #1590 @arokem
Hi @spinicist : thanks. How substantial is substantial (when do you think that you will be able to get to it)?
