### 2019-08-09 #1590 @sairamgeethanath
Here are some names/emails of potential reviewers:
1. Nikola Stikov nikola.stikov@polymtl.ca
2. Jonathan Tamir jtamir@eecs.berkeley.edu
3. Jon-Fredrick Nielsen  jfnielse@umich.edu

I am more than happy to provide a few more from the open-source Magnetic
Resonance Imaging (https://www.opensourceimaging.org/) community if need be.
