### 2019-08-09 #1637 @whedon
[REVIEW]: Virtual Scanner:  MRI on a Browser
**Submitting author:** @imr-framework (<a href="http://orcid.org/0000-0002-3776-4114">Sairam Geethanath</a>)
**Repository:** <a href="https://github.com/imr-framework/virtual-scanner" target ="_blank">https://github.com/imr-framework/virtual-scanner</a>
**Version:** 1.0.0

@nstikov, @mathieuboudreau, & @vsoch, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
