### 2019-08-09 #1590 @arokem
OK - @vsoch, @nstikov and @mathieuboudreau, please head over to #1637 for your review checklists and for the up-to-date pdf of the manuscript. Let me know if you have any questions. Thanks a lot!
