### 2019-08-09 imr-framework/virtual-scanner#21 @vsoch
JoSS Review: Manuscript
This is the start of review for https://github.com/openjournals/joss-reviews/issues/1637.

## Manuscript

The introduction does a good job to communicate a statement of need, but I'd like to see a few more references for:

 - the opening statement about MRI being excellent and having diagnostic value (a review paper would work here)
 - a reference on simulators

It would be helpful to flush out the point about vendor neutrality - most that read aren't familiar with MRI
or that there are different companies that make the machines, software, and thus simulators.

### General Comments

 - It looks like you defined "MRI" but you didnt' specify "MR" (and use it in several spots)
 - Are you able to bold the sections in the list? I almost missed the headers when I was reading. (and the same for the subsequent lists in the paper)

Very good job on explaining your roadmap, and real life use cases. This is a cool tool I'm excited to see development for.
