### 2019-08-09 imr-framework/virtual-scanner#22 @vsoch
JoSS Review: Repository
## Repository

This is for https://github.com/openjournals/joss-reviews/issues/1637

I can confirm that the repository, here, does in fact exist, and has instructions for contribution (CONTRIBUTING.md) along with a CODE_OF_CONDUCT.md. The LICENSE file is an OSI approved license, and version 1.0.0 matches what is submit with the paper.

For Authors, I can match the paper authors as follows (please correct me if I mess up)

 - Gehua Tong1: https://github.com/tonggehua
 - Sairam Geethanath1: https://github.com/sairamgeethanath
 - Marina Jimeno Manso1: https://github.com/Mmj94 ?
 - Enlin Qian1: https://github.com/qianenlin
 - Keerthi Sravan Ravi1: https://github.com/sravan953
 - Nishika Girish2
 - John Thomas Vaughan Jr.

I don't see GitHub aliases for the last 2, associated with the code - the remaining one is for the org (imr-framework) and the contributions are tiny. Could you clarify the contributions of the last two authors, if they did not work on the software directly?
