### 2019-08-09 imr-framework/virtual-scanner#23 @vsoch
JoSS Review: Functionality
## Functionality

This is for https://github.com/openjournals/joss-reviews/issues/1637

It's not required, but for the interested user it's sometimes helpful to give them a command to copy paste verbatim to clone the repository. The repository seems big, so a --depth 1 might help:

```bash
$ git clone --depth 1 https://github.com/imr-framework/virtual-scanner
$ cd virtual-scanner
```

There isn't any instruction to install dependencies, although I see the requirements.txt file, and it's immense. It would actually be sort of dangerous to just run this on the host (e.g., pip install -r requirements.txt) so I'd like to suggest that the others provide a Dockerfile that handles this. Anticipating that this is a Flask application, you could provide just a Dockerfile with an entrypoint to handle the database (sqlite?) and web server (if not development) or you could provide a docker-compose. Do any of the authors have experience with containers to support this?
