### 2019-08-09 imr-framework/virtual-scanner#22 @vsoch
Great, thanks for the clarification! Both of those contributions are very much appropriate to warrant authorship. Nishika will probably be very excited to have her name on a paper! The repository looks good for the points above, closing issue.
