### 2019-08-09 imr-framework/virtual-scanner#23 @sairamgeethanath
@vsoch We do not have experience with containers but find the review comment to be valuable. We will spend some time on this and get it to work as a Dockerfile
