### 2019-08-09 imr-framework/virtual-scanner#23 @vsoch
okay sounds good! If you give it a solid effort and it's a stopping point for the review, then I'd be happy to help - I have a lot of experience with containers.
