### 2019-08-14 #1637 @mathieuboudreau
Just as a clarifying point, I am reviewing this in tandem with @nstikov (I'm a research fellow of his), and will likely be the one most active out of both of us in this discussion/review on GitHub. This was determined in the pre-review (#1590)
