### 2019-08-14 #1637 @mathieuboudreau
@vsoch is it possible to make me able to edit the checklist status?
