### 2019-08-14 #1637 @mathieuboudreau
Ahh sorry, I didn't realize your were another reviewer, sorry about that!

No, I don't have access to Edit, as only @nstikov was assigned as the sole reviewer instead of both of us for "his" review.

@arokem can this be resolved?
