### 2019-08-14 imr-framework/virtual-scanner#24 @mathieuboudreau
JOSS REVIEW: Missing dependency (pypulseq==0.0.3)
This is for openjournals/joss-reviews#1637

Per your [installation instructions](https://github.com/imr-framework/virtual-scanner/blame/12b6a47e2fce96588adf51964474bb30540bb97d/README.md#L13), I tried to installed all your requirements (using `pip install -e .`), however I encountered a missing dependency:

```
Collecting pypulseq==0.0.3 (from virtual-scanner==1.0.21)
  ERROR: Could not find a version that satisfies the requirement pypulseq==0.0.3 (from virtual-scanner==1.0.21) (from versions: 1.0.0, 1.2.1)
ERROR: No matching distribution found for pypulseq==0.0.3 (from virtual-scanner==1.0.21)
```

It appears that this version of pypulseq is not available on pypi: https://pypi.org/project/pypulseq/#history

Can you update your requirements file, or resolve this in another way?
