### 2019-08-14 imr-framework/virtual-scanner#25 @mathieuboudreau
JOSS REVIEW: Clarifying installation details
This is for openjournals/joss-reviews#1637

In your installation instructions in your README, could you please:

1 - Specify which Python version (e.g. 3.6.x) or range of versions (e.g. 3.6-3.7) that you intend this software to be compatible with?
2 - As you are recommending people to install a lot of specific external python package & exact versions, it would be best practice to suggest to the user to install these in some sort of virtual environment so that these don't mess up their base environment (e.g. using [`conda`](https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html) or [`venv`](https://docs.python.org/3/library/venv.html), but I prefer `conda` via [miniconda](https://docs.conda.io/en/latest/miniconda.html) as it also allows for users to specify the python version of the environment).
3 - To make it easier on some folks who may be less familiar with bulk Python package installations, it would be helpful to recommend that they install all the `requirements.txt` packages using the command `pip install -e .` inside your folder, to avoid them to manually install each of them one at a time. Other solutions could work too, this is just my preferred installation workflow.
