### 2019-08-14 imr-framework/virtual-scanner#24 @mathieuboudreau
Digging into this a bit more, I see that your requirements.txt file in your master correctly suggests version 1.0.0 of pypulseq, so it's not quite clear to me why it's trying to pull version 0.0.3. I'll look into this and report back.
