### 2019-08-14 imr-framework/virtual-scanner#24 @mathieuboudreau
The conflict is in your setup.py file, which doesn't call your requirements.txt file.

https://github.com/imr-framework/virtual-scanner/blob/12b6a47e2fce96588adf51964474bb30540bb97d/setup.py#L42
