### 2019-08-14 imr-framework/virtual-scanner#24 @mathieuboudreau
There is a mismatch between your setup.py file and your requirements.txt file. At the very least, this should be cleaned up so they match, but ideally you wouldn't duplicate these.

You could either suggest that the user installs the requirements.txt file using `pip install -r requirements.txt`, or create a setup.py like [this one](https://github.com/neuropoly/axondeepseg/blob/master/setup.py), which calls the requirements.txt file on [line 36](https://github.com/neuropoly/axondeepseg/blob/ebf74c1a928ef946304d821e1e5abf0b692fd440/setup.py) using [lines 14-17](https://github.com/neuropoly/axondeepseg/blob/ebf74c1a928ef946304d821e1e5abf0b692fd440/setup.py#L14-L17).
