### 2019-08-14 imr-framework/virtual-scanner#24 @mathieuboudreau
Note that this bug is present while installing using pypi  using `pip install virtual-scanner`, so I would recommend using the setup.py solution I recommend above if you plan on hosting on pypi.
