### 2019-08-14 imr-framework/virtual-scanner#26 @mathieuboudreau
JOSS REVIEW: Matplotlib issue on Mac when using (Ana)conda
This is for openjournals/joss-reviews#1637

Not a high priority, but it may be worth noting to your users that if they're running on a Mac with (Ana)conda, the fix if they encounter the following error when running `python virtualscanner/coms/coms_ui/coms_server_flask.py`:

>ImportError: Python is not installed as a framework. The Mac OS X backend will not be able to function correctly if Python is not installed as a framework. See the Python documentation for more information on installing Python as a framework on Mac OS X. Please either reinstall Python as a framework, or try one of the other backends. If you are using (Ana)Conda please install python.app and replace the use of 'python' with 'pythonw'. See 'Working with Matplotlib on OSX' in the Matplotlib FAQ for more information.

is to run this command:

`echo "backend: TkAgg" >> ~/.matplotlib/matplotlibrc`

See: https://github.com/scikit-optimize/scikit-optimize/issues/637#issuecomment-369730420
