### 2019-08-14 #1637 @vsoch
@mathieuboudreau here is the markdown https://gist.github.com/vsoch/c9f65143293acb975af251608b314579

Just view the raw, make changes somewhere that I can view raw (another gist) and I'll copy paste it in for you.
