### 2019-08-14 imr-framework/virtual-scanner#27 @mathieuboudreau
JOSS REVIEW: Running exam results in ModuleNotFoundError: No module named 'virtualscanner'
This is for openjournals/joss-reviews#1637

Here's my setup:

* anaconda already installed on my system
* cloned the repo and `cd`' into it
* created a virtual environment using `conda create -n virtualscanner python=3.6`
* activated the virtual environment using `conda activate virtualscanner`
* installed requirements using 'pip install -r requirements.txt`
* Resolved the Matplotlib/Conda/Mac issue using `echo "backend: TkAgg" >> ~/.matplotlib/matplotlibrc`
* loaded the server using `python virtualscanner/coms/coms_ui/coms_server_flask.py`
* Opened the page in a browser: http://0.0.0.0:5000/
* "logged" in with my email and began scan
* Registered "numerical" subject
* Selected "GRE" sequence
* Clicked "Run exam"

Here, the console did nothing, and the terminal gave the following logged output:

```

(virtualscanner) mathieuboudreau@brassens:~/neuropoly/JOSS/virtual-scanner$ python virtualscanner/coms/coms_ui/coms_server_flask.py
Using TensorFlow backend.
 * Serving Flask app "coms_server_flask" (lazy loading)
 * Environment: production
   WARNING: Do not use the development server in a production environment.
   Use a production WSGI server instead.
 * Debug mode: on
 * Running on http://0.0.0.0:5000/ (Press CTRL+C to quit)
 * Restarting with stat
Using TensorFlow backend.
 * Debugger is active!
 * Debugger PIN: 160-022-916
127.0.0.1 - - [14/Aug/2019 15:27:34] "GET / HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:27:34] "GET /static/packages/jquery-3.3.1.min.js HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:27:34] "GET /static/logo-blue-white-onlycrown.png HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:27:34] "GET /static/log_in_style.css HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:27:34] "GET /static/log_in_func.js HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:27:34] "GET /static/main_style.css HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:27:34] "GET /static/virtual_scanner_logo_v1.png HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:27:35] "GET /static/favicon.ico HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:32:12] "POST / HTTP/1.1" 302 -
127.0.0.1 - - [14/Aug/2019 15:32:12] "GET /register HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:32:12] "GET /static/packages/jQuery-ui/jquery-ui.css HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:32:12] "GET /static/packages/Bootstrap/bootstrap.min.css HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:32:12] "GET /static/packages/jQuery-ui/jquery-ui.js HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:32:12] "GET /static/packages/Bootstrap/popper.min.js HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:32:12] "GET /static/packages/Bootstrap/jquery-3.3.1.slim.min.js HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:32:12] "GET /static/packages/Bootstrap/bootstrap.min.js HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:32:12] "GET /static/phantomAutofill.js HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:32:12] "GET /static/register_style.css HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:32:12] "GET /static/register_func.js HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:32:12] "GET /static/log_out_func.js HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:32:17] "GET /static/phantom_pics/Numerical.png HTTP/1.1" 200 -
{'formName': 'reg', 'subjecttype': 'Numerical', 'patid': '3689', 'name': 'Numerical', 'age': '00', 'dob': '8/14/2019', 'gender': 'other', 'weight': '3', 'weight-unit': 'kg', 'height': '20', 'height-unit': 'cm', 'inches': '', 'orient': 'HFS', 'anatomy': 'brain'}
0
127.0.0.1 - - [14/Aug/2019 15:32:24] "POST /receiver HTTP/1.1" 302 -
127.0.0.1 - - [14/Aug/2019 15:32:24] "GET /register_success HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:32:26] "GET /acquire HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:32:26] "GET /static/acquire_func.js HTTP/1.1" 200 -
127.0.0.1 - - [14/Aug/2019 15:32:26] "GET /static/acquire_style.css HTTP/1.1" 200 -
[('Numerical', 3689, 'Numerical', 0, '8/14/2019', 'other', 3.0, 20.0, 'HFS', 'brain')]
{'formName': 'acq', 'selectedSeq': 'GRE', 'TR': '200', 'TE': '10', 'FA': '90', 'sl-orient': 'axial', 'thck': '16', 'slicenum': '1', 'freq': 'x', 'ph': 'y', 'Nx': '15', 'Ny': '15', 'FOVx': '240', 'FOVy': '240'}
Traceback (most recent call last):
  File "/Users/mathieuboudreau/neuropoly/joss/virtual-scanner/virtualscanner/server/simulation/bloch/pulseq_bloch_simulator.py", line 13, in <module>
    import virtualscanner.server.simulation.bloch.pulseq_blochsim_methods as blcsim
ModuleNotFoundError: No module named 'virtualscanner'
[('Numerical', 3689, 'Numerical', 0, '8/14/2019', 'other', 3.0, 20.0, 'HFS', 'brain')]
{'formName': 'acq', 'selectedSeq': 'GRE', 'TR': '200', 'TE': '10', 'FA': '90', 'sl-orient': 'axial', 'thck': '16', 'slicenum': '1', 'freq': 'x', 'ph': 'y', 'Nx': '15', 'Ny': '15', 'FOVx': '240', 'FOVy': '240'}
Traceback (most recent call last):
  File "/Users/mathieuboudreau/neuropoly/joss/virtual-scanner/virtualscanner/server/simulation/bloch/pulseq_bloch_simulator.py", line 13, in <module>
    import virtualscanner.server.simulation.bloch.pulseq_blochsim_methods as blcsim
ModuleNotFoundError: No module named 'virtualscanner'
[('Numerical', 3689, 'Numerical', 0, '8/14/2019', 'other', 3.0, 20.0, 'HFS', 'brain')]
{'formName': 'acq', 'selectedSeq': 'GRE', 'TR': '200', 'TE': '10', 'FA': '90', 'sl-orient': 'axial', 'thck': '16', 'slicenum': '1', 'freq': 'x', 'ph': 'y', 'Nx': '15', 'Ny': '15', 'FOVx': '240', 'FOVy': '240'}
Traceback (most recent call last):
  File "/Users/mathieuboudreau/neuropoly/joss/virtual-scanner/virtualscanner/server/simulation/bloch/pulseq_bloch_simulator.py", line 13, in <module>
    import virtualscanner.server.simulation.bloch.pulseq_blochsim_methods as blcsim
ModuleNotFoundError: No module named 'virtualscanner'

```

Strangely, if I opened a `python` session and tried to import `import virtualscanner.server.simulation.bloch.pulseq_blochsim_methods as blcsim`, it imported without any issues. So I'm wondering if this is an issue with the flask server not calling these in my virtual environment? Maybe it's running a python session on my base python environment, but I don't know enough about flask/web to know. If that's the case, I would suggest to either 1) find a solution for this (recommended), or 2) notify the users to not run this software on a virtual environment (not ideal).
