### 2019-08-14 imr-framework/virtual-scanner#27 @sairamgeethanath
@sravan953 Seems to be a path issue, does it need an edit in the constants.py?
