### 2019-08-14 imr-framework/virtual-scanner#25 @sairamgeethanath
@mathieuboudreau We will update the install instructions but we use Python 3.6 and [venv](https://docs.python.org/3/library/venv.html). I request you to use the same for reproducible results. You can pip install the package as well with pip install virtual-scanner
