### 2019-08-14 imr-framework/virtual-scanner#25 @mathieuboudreau
Thanks! I did use 3.6 (but prefer conda over venv, as venv will only use your default python version and can't set it to a different one, and my default one is 3.7)

 However just to let you know that I encountered the issue #24 in when trying to install using `pip install virtual-scanner`, so you may want to investigate that.
