### 2019-08-14 imr-framework/virtual-scanner#26 @sravan953
@sairamgeethanath This should go into the documentation under a 'Common errors' section?
