### 2019-08-14 f3718ac
Update setup.py to fix dependencies
Install dependencies from requirements.txt
