### 2019-08-14 b42f909
Update installation instructions in README
