### 2019-08-14 imr-framework/virtual-scanner#25 @sravan953
@mathieuboudreau Thank you for your inputs, they've been addressed in b42f9099812e2024ab4541c47f5b4a23b040a500.
