+++ virtualscanner/server/simulation/bloch/spingroup_ps.py
@@ -1,8 +1,10 @@
 # Copyright of the Board of Trustees of Columbia University in the City of New York
 
 import numpy as np
-GAMMA = 2*42.58e6 * np.pi
+
+GAMMA = 2 * 42.58e6 * np.pi
 GAMMA_BAR = 42.58e6
+
 
 class SpinGroup:
     """Basic magnetization unit for Bloch simulation
@@ -44,14 +46,14 @@ class SpinGroup:
 
     """
 
-    def __init__(self, loc=(0,0,0), pdt1t2=(1,0,0), df=0):
+    def __init__(self, loc=(0, 0, 0), pdt1t2=(1, 0, 0), df=0):
         self.m = np.array([[0], [0], [1]])
         self.PD = pdt1t2[0]
         self.T1 = pdt1t2[1]
         self.T2 = pdt1t2[2]
         self.loc = loc
         self.df = df
-        self.signal=[]
+        self.signal = []
 
     def get_m_signal(self):
         """Gets spin group's transverse magnetization
@@ -64,11 +66,11 @@ class SpinGroup:
             Imaginary part = My
 
         """
-        m_signal = np.squeeze(self.PD*(self.m[0] + 1j * self.m[1]))
+        m_signal = np.squeeze(self.PD * (self.m[0] + 1j * self.m[1]))
 
         return m_signal
 
-    def fpwg(self,grad_area,t):
+    def fpwg(self, grad_area, t):
         """Apply only gradients to the spin group
 
         The spin group precesses with gradients applied.
@@ -83,15 +85,15 @@ class SpinGroup:
             Total time of precession in seconds
 
         """
-        x,y,z = self.loc
-        phi = GAMMA*(x*grad_area[0]+y*grad_area[1]+z*grad_area[2])+2*np.pi*self.df*t
+        x, y, z = self.loc
+        phi = GAMMA * (x * grad_area[0] + y * grad_area[1] + z * grad_area[2]) + 2 * np.pi * self.df * t
         C, S = np.cos(phi), np.sin(phi)
-        E1 = 1 if self.T1 == 0 else np.exp(-t/self.T1)
-        E2 = 1 if self.T2 == 0 else np.exp(-t/self.T2)
-        A = np.array([[E2*C, E2*S, 0],
-                      [-E2*S, E2*C, 0],
+        E1 = 1 if self.T1 == 0 else np.exp(-t / self.T1)
+        E2 = 1 if self.T2 == 0 else np.exp(-t / self.T2)
+        A = np.array([[E2 * C, E2 * S, 0],
+                      [-E2 * S, E2 * C, 0],
                       [0, 0, E1]])
-        self.m = A@self.m + [[0],[0],[1 - E1]]
+        self.m = A @ self.m + [[0], [0], [1 - E1]]
 
     def delay(self, t):
         """Applies a time passage to the spin group
@@ -105,17 +107,15 @@ class SpinGroup:
             Delay interval in seconds
 
         """
-        self.T1 = max(0,self.T1)
-        self.T2 = max(0,self.T2)
-        E1 = 1 if self.T1 == 0 else np.exp(-t/self.T1)
-        E2 = 1 if self.T2 == 0 else np.exp(-t/self.T2)
+        self.T1 = max(0, self.T1)
+        self.T2 = max(0, self.T2)
+        E1 = 1 if self.T1 == 0 else np.exp(-t / self.T1)
+        E2 = 1 if self.T2 == 0 else np.exp(-t / self.T2)
 
         A = np.array([[E2, E2, 0],
                       [-E2, E2, 0],
                       [0, 0, E1]])
-        self.m = A@self.m + np.array([[0], [0], [1 - E1]])
-
-
[…]
