+++ virtualscanner/server/simulation/bloch/pulseq_blochsim_methods.py
@@ -3,18 +3,14 @@
 Methods to help Bloch simulation from pulseq objects
 """
 
-import numpy as np
-import numpy.matlib as npm
-import matplotlib.pyplot as plt
-import time
-import virtualscanner.server.simulation.bloch.phantom as pht
-import multiprocessing as mp
-import virtualscanner.server.simulation.bloch.spingroup_ps as sg
-import virtualscanner.server.simulation.bloch.pulseq_library as psl
 from math import pi
 
+import numpy as np
+
+import virtualscanner.server.simulation.bloch.spingroup_ps as sg
+
 GAMMA_BAR = 42.5775e6
-GAMMA = 2*pi*GAMMA_BAR
+GAMMA = 2 * pi * GAMMA_BAR
 
 
 def store_pulseq_commands(seq):
@@ -51,9 +47,9 @@ def store_pulseq_commands(seq):
             commands += 'p'
             rf_time = np.array(this_blk['rf'].t[0]) - dt_rf
             df = this_blk['rf'].freq_offset
-            b1 = np.multiply(np.exp(-2*pi*1j*df*rf_time),this_blk['rf'].signal/GAMMA_BAR)
+            b1 = np.multiply(np.exp(-2 * pi * 1j * df * rf_time), this_blk['rf'].signal / GAMMA_BAR)
             rf_grad, rf_timing, rf_duration = combine_gradients(blk=this_blk, timing=rf_time)
-            seq_params.append([b1,rf_grad,dt_rf])
+            seq_params.append([b1, rf_grad, dt_rf])
 
         # Case 3: ADC sampling
         elif event_row[5] != 0:
@@ -62,21 +58,21 @@ def store_pulseq_commands(seq):
             dt_adc = adc.dwell
             delay = adc.delay
             grad, timing, duration = combine_gradients(blk=this_blk, dt=dt_adc, delay=delay)
-            seq_params.append([dt_adc,int(adc.num_samples),delay,grad,timing])
+            seq_params.append([dt_adc, int(adc.num_samples), delay, grad, timing])
 
         # Case 4: just gradients
         elif event_row[2] != 0 or event_row[3] != 0 or event_row[4] != 0:
             commands += 'g'
             # Process gradients
             fp_grads_area = combine_gradient_areas(blk=this_blk)
-            dur = find_precessing_time(blk=this_blk,dt=dt_grad)
-            seq_params.append([fp_grads_area,dur])
+            dur = find_precessing_time(blk=this_blk, dt=dt_grad)
+            seq_params.append([fp_grads_area, dur])
 
-    seq_info = {'commands':commands, 'params':seq_params,'grad_raster_time':dt_grad}
+    seq_info = {'commands': commands, 'params': seq_params, 'grad_raster_time': dt_grad}
     return seq_info
 
 
-def apply_pulseq_commands(isc,seq_info):
+def apply_pulseq_commands(isc, seq_info):
     """Imposes sequence commands on a single spin group
 
     This is the key simulation function that goes through the commands and applies each to the spin group
@@ -94,17 +90,17 @@ def apply_pulseq_commands(isc,seq_info):
     for c in range(len(cmds)):
         cstr = cmds[c]
         cpars = pars[c]
-        if cstr == 'd': # delay
+        if cstr == 'd':  # delay
             isc.delay(t=cpars[0])
-        elif cstr == 'p': # rf pulse
-            isc.apply_rf(pulse_shape=cpars[0],grads_shape=cpars[1],dt=cpars[2])
-        elif cstr == 'r': # Readout
-            isc.readout(dwell=cpars[0],n=cpars[1],delay=cpars[2],grad=cpars[3],timing=cpars[4])
[…]

[…]
