+++ virtualscanner/server/simulation/bloch/pulseq_bloch_simulator.py
@@ -10,6 +10,13 @@ Notes
 See beginning of the main section for explanations of the command line arguments (line 41-71)
 
 """
+if __name__ == '__main__':
+    import os
+    import sys
+
+    script_path = os.path.abspath(__file__)
+    SEARCH_PATH = script_path[:script_path.index('virtual-scanner') + len('virtual-scanner') + 1]
+    sys.path.insert(0, SEARCH_PATH)
 
 import matplotlib
 import numpy as np
@@ -17,6 +24,7 @@ import numpy as np
 import virtualscanner.server.simulation.bloch.pulseq_blochsim_methods as blcsim
 
 matplotlib.use('Agg')
+import matplotlib.pyplot as pltw
 import time
 import virtualscanner.server.simulation.bloch.phantom as pht
 import multiprocessing as mp
