### 2019-08-14 imr-framework/virtual-scanner#26 @sairamgeethanath
@sravan953 @mathieuboudreau If we all agree then we can file this in known issues as this is a package (matplotlib) dependency issue.
