### 2019-08-14 imr-framework/virtual-scanner#26 @mathieuboudreau
That works for me!
