### 2019-08-15 a2433fa
Update freq and phase values wrt sl_orient from previous results
Fix #18
