+++ virtualscanner/coms/coms_ui/static/acquire_func.js
@@ -195,12 +195,36 @@ function autoFillForm(dict){
     $("#TI").val(dict['TI']);
   }
   $("#sl-orient").val(dict['sl-orient']);
+  var slorient = dict['sl-orient'];
+  if (slorient == "axial"){
+    $(".freq#x").prop("disabled",false);
+    $("#freq").val("x");
+    $(".phase#y").prop("disabled",false);
+    $("#phase").val("y");
+    $(".freq:not(#x)").prop("disabled",true);
+    $(".phase:not(#y)").prop("disabled",true);
+  }
+  else if (slorient == "sagittal") {
+    $(".freq#y").prop("disabled",false);
+    $("#freq").val("y");
+    $(".phase#z").prop("disabled",false);
+    $("#phase").val("z");
+    $(".freq:not(#y)").prop("disabled",true);
+    $(".phase:not(#z)").prop("disabled",true);
+  }
+  else if (slorient == "coronal") {
+    $(".freq#z").prop("disabled",false);
+    $("#freq").val("z");
+    $(".phase#x").prop("disabled",false);
+    $("#phase").val("x");
+    $(".freq:not(#z)").prop("disabled",true);
+    $(".phase:not(#x)").prop("disabled",true);
+  }
   $("#thickness").val(dict['thck']);
   $("#slices").val(dict['slicenum']);
   $("#sl-gap").prop("disabled",true);
   $("#sl-gap").css("background-color","#bfbfbf");
-  $("#freq").val(dict['freq']);
-  $("#phase").val(dict['ph']);
+
   $("#ADC-bw").val(dict['bw']);
   $("#Nx").val(dict['Nx']);
   $("#Ny").val(dict['Ny']);
