### 2019-08-15 0e805c7
Correction of line leading to image display errors
Fix #19 Fix #20
