### 2019-08-19 imr-framework/virtual-scanner#29 @mathieuboudreau
JOSS REVIEW: Clarifying how to start software
This is for openjournals/joss-reviews#1637

To me, the "Getting started" instructions appear to lack a bit of clarity when it comes to using the software (both from pipy installation and from cloning the repo), as I encountered issues when following your instructions explicitly.

pypi instructions
* After installing with pipy, you're [Getting started]() instructions simply state to "run coms_server_flask.py". To people with Python experience, running a *.py file would mean to type the command `python coms_server_flask.py` in their terminal, however this fails as coms_server_flask.py is not in the path. It's not easy for users to find the correct full path of this file when it's in a virtual environment.

cloned repo instructions
* Same issue as above, except now since the user sees the code, they can dig in themselves. Still not every elegant or clear for a quick start.

I think both of these could be resolved using the [entrypoint](https://github.com/imr-framework/virtual-scanner/blob/0e805c74cd780862d617de2437eb9d54375339cd/setup.py#L27-L30) that's already implemented, but not described in the documentation. Two minor changes could solve this:

1. On [this line](https://github.com/imr-framework/virtual-scanner/blame/0e805c74cd780862d617de2437eb9d54375339cd/README.md#L21) of your README, change the command to `pip install -e .` so that the **setup.py** file is executed instead of only the requirements file, which will then create the command-line entrypoint and add it to the path.

2. On [this line](https://github.com/imr-framework/virtual-scanner/blame/0e805c74cd780862d617de2437eb9d54375339cd/README.md#L23) of the README file, change the command to simply `virtualscanner`.
