### 2019-08-19 imr-framework/virtual-scanner#30 @mathieuboudreau
JOSS REVIEW: Affiliation 1
This is for openjournals/joss-reviews#1637

In your first affiliation, would you mind adding the state and country at the end to be consistent with your second affiliation?
