### 2019-08-19 imr-framework/virtual-scanner#30 @sairamgeethanath
@tonggehua Can you please take a crack at all manuscript related fixes including this one? Let's schedule a Friday submission of the paper.md
