### 2019-08-19 imr-framework/virtual-scanner#29 @mathieuboudreau
Actually, if you go this route, you'll need a third change, as this section of the code is only run if you call `coms_server_flask.py` as through main, and you need `os` for sure.

https://github.com/imr-framework/virtual-scanner/blob/0e805c74cd780862d617de2437eb9d54375339cd/virtualscanner/coms/coms_ui/coms_server_flask.py#L4-L10

This if statement needs to be refactored somehow so that it's applied (or at least, `os` is imported) if/when `launch_virtualscanner()` is called
