### 2019-08-19 imr-framework/virtual-scanner#30 @tonggehua
Understood!

Gehua

On Mon, Aug 19, 2019 at 2:32 PM Sairam Geethanath <notifications@github.com>
wrote:

> @tonggehua <https://github.com/tonggehua> Can you please take a crack at
> all manuscript related fixes including this one? Let's schedule a Friday
> submission of the paper.md
>
