### 2019-08-20 imr-framework/virtual-scanner#26 @mathieuboudreau
FYI: I found another solution that works better for me: https://matplotlib.org/3.1.0/faq/osx_framework.html

* Install python.app (`conda install python.app`) and use `pythonw` rather than `python`.
