### 2019-08-20 imr-framework/virtual-scanner#32 @mathieuboudreau
JOSS REVIEW: Statement of need (README.md)
This is for openjournals/joss-reviews#1637

Would you mind expanding your README.md a little bit more so that it adheres to this JoSS review checklist point?

> **A statement of need**: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

You link to the Africa Challenge in the ISMRM website, which I think states the target audience and the problem your software tries to resolve, however maybe you could briefly explain that in your own words here (in part, because the page you linked to on the ISMRM website might be broken sometime in the future). Just 2-3 sentences would suffice.
