### 2019-08-20 imr-framework/virtual-scanner#33 @mathieuboudreau
JOSS REVIEW: Community guidelines
This is for openjournals/joss-reviews#1637

Could you expand your [CONTRIBUTING.md](https://github.com/imr-framework/virtual-scanner/blob/master/CONTRIBUTING.md) a bit so that it has sections describing the three points required by the JoSS checklist item?

>**Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Basically indicating what your preference is for people to contribute (e.g. Fork and make a pull request?), your preference for how to report problems (GitHub issues or email?), and how to reach out for support (issues, email, or any other method of your preference)?

These can be very brief. Also, for folks not familiar with the CONTRIBUTING.md file, maybe make a quick mention of it in your README.md and link to it.
