### 2019-08-20 fa324a8
Update Statement of Need in response to #32
