### 2019-08-22 1fb83df
Update CONTRIBUTING.md in accord to #33
