### 2019-08-23 035c12a
Add KNOWN_ISSUES.md
- Reference KNOWN_ISSUES.md in README.md
- Minor formatting changes
