### 2019-08-23 0d6740b
Fix KNOWN_ISSUES.md link
