### 2019-08-23 b1de927
Fix KNOWN_ISSUES.md
