### 2019-08-23 3de4799
Fix KNOWN_ISSUES.md
