### 2019-08-23 2559589
Linking to CONTRIBUTING.md #33
