+++ README.md
@@ -45,4 +45,11 @@ Alternatively, you can read the API documentation [here](https://imr-framework.g
 * Other features, including phantom and sequence viewers and reconstruction methods, are in active development.
 
 ## Known Issues
-Please refer to the [Known Issues](https://github.com/imr-framework/virtual-scanner/blob/master/KNOWN_ISSUES.md) document.
+Please refer to the [Known Issues](known-issues) document.
+
+[known-issues]: https://github.com/imr-framework/virtual-scanner/blob/master/KNOWN_ISSUES.md
+
+## Contributing
+If you would like to contribute to Virtual Scanner, please take a look at [CONTRIBUTING.md](contributing).
+
+[contributing]: https://github.com/imr-framework/virtual-scanner/blob/master/CONTRIBUTING.md
