### 2019-08-23 79bdeb6
Linking to CONTRIBUTING.md #33
