### 2019-08-23 imr-framework/virtual-scanner#33 @tonggehua
Close with commit b163a1b
