+++ README.md
@@ -45,11 +45,7 @@ Alternatively, you can read the API documentation [here](https://imr-framework.g
 * Other features, including phantom and sequence viewers and reconstruction methods, are in active development.
 
 ## Known Issues
-Please refer to the [Known Issues](known-issues) document.
-
-[known-issues]: https://github.com/imr-framework/virtual-scanner/blob/master/KNOWN_ISSUES.md
+Please refer to the [Known Issues](https://github.com/imr-framework/virtual-scanner/blob/master/KNOWN_ISSUES.md) document.
 
 ## Contributing
-If you would like to contribute to Virtual Scanner, please take a look at [CONTRIBUTING.md](contributing).
-
-[contributing]: https://github.com/imr-framework/virtual-scanner/blob/master/CONTRIBUTING.md
+If you would like to contribute to Virtual Scanner, please take a look at the [Community Guidelines](https://github.com/imr-framework/virtual-scanner/blob/master/CONTRIBUTING.md) document.
