### 2019-08-23 1579542
Update README.md
Edits to the statement of need - refinement only
