### 2019-08-23 imr-framework/virtual-scanner#26 @sravan953
Referenced this issue in the `KNOWN_ISSUES.md`, commit 035c12ae631eace096858bb516dc90e81c39eba5.
