### 2019-08-23 imr-framework/virtual-scanner#32 @tonggehua
Close with fa324a8
