### 2019-08-23 50d385d
Close #21
Manuscript changes and citation additions
