+++ paper/paper.bib
@@ -159,3 +159,49 @@ doi={10.1126/science.1925560}
   publisher={IOP Publishing},
   doi={10.1088/0022-3719/10/3/004}
 }
+
+@article{webb2003introduction,
+  title={Introduction to biomedical imaging},
+  author={Webb, Andrew and Kagadis, George C},
+  journal={Medical Physics},
+  volume={30},
+  number={8},
+  chapter={4},
+  pages={210-214},
+  year={2003},
+  publisher={Wiley Online Library}
+}
+
+@Article{Rao2015,
+author={Rao, Rashmi
+and Venkatesan, Ramesh
+and Geethanath, Sairam},
+title={Role of MRI in medical diagnostics},
+journal={Resonance},
+year={2015},
+month={Nov},
+day={01},
+volume={20},
+number={11},
+pages={1003--1011},
+issn={0973-712X},
+doi={10.1007/s12045-015-0268-2},
+url={https://doi.org/10.1007/s12045-015-0268-2}
+}
+
+@book{mcrobbie2017mri,
+  title={MRI from Picture to Proton},
+  author={McRobbie, Donald W and Moore, Elizabeth A and Graves, Martin J and Prince, Martin R},
+  year={2017},
+  publisher={Cambridge university press},
+  chapter={2}
+}
+
+@book{brown2015mri,
+  title={MRI: basic principles and applications},
+  author={Brown, Mark A and Semelka, Richard C and Dale, Brian M},
+  year={2015},
+  publisher={John Wiley \& Sons},
+  chapter={17}
+}
+
