+++ paper/paper.md
@@ -30,63 +30,63 @@ authors:
     affiliation: 1
 
 affiliations:
- - name: Columbia Magnetic Resonance Research Center, Columbia University in the City of New York
+ - name: Columbia Magnetic Resonance Research Center, Columbia University in the City of New York, New York, NY, USA
    index: 1
  - name: Plum Grove Junior High, Palatine, IL, USA
    index: 2
  
-date: 26 July 2019
+date: 23 August 2019
 
 bibliography: paper.bib
 ---
 # Summary
-Magnetic Resonance Imaging (MRI) is a medical imaging modality that provides excellent soft-tissue contrast with high diagnostic value. 
-System simulators are key to method development in MRI because of its inherent complexity. However, most existing MR system-wide simulators are proprietary to manufacturing companies, and an open-source, integrated webtool would help MR researchers around the world develop and share new methodology. Specifically, it would provide vendor-neutral tools for rapidly prototyping MR acquisition methods and hardware components, services for sharing computational methods and resources, and education for under-served areas to develop their own MR expertise [@obungoloch_2019]. Being vendor-neutral would also enable multi-site repeatability studies. To fulfill this need, we designed ``Virtual Scanner`` with three main characteristics in mind:
+Magnetic Resonance Imaging (MRI) is a medical imaging modality that provides excellent soft tissue contrast with high diagnostic value [@webb2003introduction; @Rao2015; @mcrobbie2017mri; @brown2015mri]. System simulators are important because of the inherent complexity of MRI. However, most existing MRI simulators are  proprietary to manufacturing companies. In addition, typical existing open-source simulators do not support both system-wide simulation and pulse sequence deployment on multiple vendors (Supplementary table S1, @ravi2018pulseq). An open-source, integrated, and vendor-neutral webtool will help MRI researchers around the world develop and share new methodology. Vendor-neutrality allows imaging methods to be applied on scanners from multiple companies and streamlines multi-site repeatability studies. This software should provide functionalities for rapidly prototyping MR methods, services for sharing computational resources, and educational features for under-served areas to develop their own MRI expertise [@obungoloch_2019].
 
+To fulfill this need, we designed ``Virtual Scanner`` with three main characteristics in mind:
 
-* Modular: ``Virtual Scanner`` consists of modules corresponding to steps in the MR signal chain [@webb2016magnetic]. Since modern MRI systems contain multiple levels of hardware and software, it is important for researchers to select the appropriate steps for simulation when developing acquisition paradigms, analysis methods, or hardware components. Students may also choose to focus on individual aspects of the experiment.
+* Modular: ``Virtual Scanner`` consists of modules corresponding to steps in the MRI signal chain [@webb2016magnetic]. Since modern MRI systems contain multiple levels of hardware and software, it is important for researchers to select the appropriate steps for simulation when developing acquisition paradigms, analysis methods, or hardware components. Students may also choose to focus on individual aspects of the experiment.
 
 * Zero-footprint: ``Virtual Scanner`` was implemented as a web application using the FLASK framework [@flask] with backend code in Python. The local server can be set up by running a script or using the command line. Once the server is up, one can access it on a browser without additional steps for installing any software packages. 
 
-* Open-source: existing tools in the MR community are incorporated for effective creation and sharing of resources. Pulseq, a multi-vendor MR pulse sequence format [@layton2017pulseq; @ravi2018pulseq], can be directly fed into the simulator. Since translating from simulation to experiments on a real scanner is a key step in MR methods development, being able to deploy the same sequence file can help streamline this process. 
[…]
