### 2019-08-23 imr-framework/virtual-scanner#21 @vsoch
Sorry, could you please explain how the issue has been addressed?
