### 2019-08-23 b7a5c58
Update ROI detection method
