### 2019-08-23 imr-framework/virtual-scanner#21 @tonggehua
1. We added 4 references that discuss the diagnostic uses of MRI.
2. We added a reference on open-source simulators (in Supporting table 1 in the Pulseq-GPI paper). However, we couldn't find references to proprietary simulators.
3. All the "MR"s have been changed to "MRI"s to avoid confusion.
4. We didn't change the styling since the main headers are already bolded in the PDF (https://github.com/openjournals/joss-papers/blob/joss.01637/joss.01637/10.21105.joss.01637.pdf). If you would prefer the first word after the bullet points to be bold, I can do that.
