### 2019-08-23 imr-framework/virtual-scanner#21 @vsoch
Thank you @tonggehua ! I still have most of my review to do - I'll try to do it within the next week (a lot on my plate at the moment!)
