### 2019-08-23 44dbb5d
Update installation instructions in README.md
