### 2019-08-23 imr-framework/virtual-scanner#29 @sravan953
@mathieuboudreau I have updated the `README.md` based on your inputs. Please let us know if this works and we can go ahead and close the issue.
