### 2019-08-23 c68d31c
Fix #34. Disable Run Exam button when no sequence is selected
