### 2019-08-23 f8ecdaf
Fix OS inconsistency in utest_ROI_analysis_T1_T2.py
