+++ virtualscanner/server/ana/utest_ROI_analysis_T1_T2.py
@@ -1,13 +1,18 @@
+import sys
 import unittest
-import virtualscanner.server.ana.ROI_analysis as dicomROIanalysis
+
 import numpy as np
 from scipy.misc import imread
-from virtualscanner.utils import constants
-import sys
 
-SERVER_T1_MAP_PATH = constants.SERVER_ANALYZE_PATH / 'inputs' / 'T1_map'
-SERVER_T2_MAP_PATH = constants.SERVER_ANALYZE_PATH / 'inputs' / 'T2_map'
+import virtualscanner.server.ana.ROI_analysis as dicomROIanalysis
+from virtualscanner.utils import constants
+
+SERVER_T1_WINDOWS_MAP_PATH = constants.SERVER_ANALYZE_PATH / 'inputs' / 'T1_map_windows'
+SERVER_T1_MAC_MAP_PATH = constants.SERVER_ANALYZE_PATH / 'inputs' / 'T1_map_mac'
+SERVER_T2_WINDOWS_MAP_PATH = constants.SERVER_ANALYZE_PATH / 'inputs' / 'T2_map_windows'
+SERVER_T2_MAC_MAP_PATH = constants.SERVER_ANALYZE_PATH / 'inputs' / 'T2_map_mac'
 COMS_MAP_PATH = constants.COMS_UI_PATH / 'static' / 'ana' / 'outputs' / '9306'
+
 
 class MyTestCase(unittest.TestCase):
 
@@ -15,13 +20,13 @@ class MyTestCase(unittest.TestCase):
         """
         Unit test ROI analysis of T1 map.
         """
-        platform = {'win32': 'utest_T1_map_with_ROI_windows.png',
-                    'darwin': 'utest_T1_map_with_ROI_mac.png'}
-        filename = platform[sys.platform]
-        roi_filename = dicomROIanalysis.main(dicom_map_path=SERVER_T1_MAP_PATH, map_type='T1', map_size='128',
-                                                fov='170', pat_id='9306')
+        platform = {'win32': SERVER_T1_WINDOWS_MAP_PATH,
+                    'darwin': SERVER_T1_MAC_MAP_PATH}
+        dicom_map_path = platform[sys.platform]
+        roi_filename = dicomROIanalysis.main(dicom_map_path=dicom_map_path, map_type='T1', map_size='128',
+                                             fov='170', pat_id='9306')
         generated_ROI = imread(COMS_MAP_PATH / roi_filename)
-        utest_ROI = imread(SERVER_T1_MAP_PATH / filename)
+        utest_ROI = imread(dicom_map_path / 'utest_T1_map_with_ROI.png')
 
         np.testing.assert_allclose(generated_ROI, utest_ROI)
 
@@ -29,13 +34,13 @@ class MyTestCase(unittest.TestCase):
         """
         Unit test ROI analysis of T2 map.
         """
-        platform = {'win32': 'utest_T2_map_with_ROI_windows.png',
-                    'darwin': 'utest_T2_map_with_ROI_mac.png'}
-        filename = platform[sys.platform]
-        roi_filename = dicomROIanalysis.main(dicom_map_path=SERVER_T2_MAP_PATH, map_type='T2', map_size='128',
-                                                fov='210', pat_id='9306')
+        platform = {'win32': SERVER_T2_WINDOWS_MAP_PATH,
+                    'darwin': SERVER_T2_MAC_MAP_PATH}
+        dicom_map_path = platform[sys.platform]
+        roi_filename = dicomROIanalysis.main(dicom_map_path=dicom_map_path, map_type='T2', map_size='128',
+                                             fov='210', pat_id='9306')
[…]
