### 2019-08-23 bdcc585
Fix ROI analysis on Docker
