### 2019-08-24 20f0365
Fix imports for console launch
`os` module is always imported
