### 2019-08-24 imr-framework/virtual-scanner#21 @vsoch
Okay, just checking over the points above now! For the bolding, I am referring to the first word (before the colon) after each section. For example, we would change:

```markdown
* Modular: ``Virtual Scanner`` consists of modules corresponding to steps in the MRI signal chain [@webb2016magnetic]. Since modern MRI systems contain multiple levels of hardware and software, it is important for researchers to select the appropriate steps for simulation when developing acquisition paradigms, analysis methods, or hardware components. Students may also choose to focus on individual aspects of the experiment.
```
to
```markdown
* **Modular**: ``Virtual Scanner`` consists of modules corresponding to steps in the MRI signal chain [@webb2016magnetic]. Since modern MRI systems contain multiple levels of hardware and software, it is important for researchers to select the appropriate steps for simulation when developing acquisition paradigms, analysis methods, or hardware components. Students may also choose to focus on individual aspects of the experiment.
```
And the difference in markdown (rendered here) is:

* Modular: ``Virtual Scanner`` consists of modules corresponding to steps in the MRI signal chain
* **Modular**: ``Virtual Scanner`` consists of modules corresponding to steps in the MRI signal chain

Given how it renders, I suggested this change because I missed that there was a small section heading with a colon when I first read. Does that make sense?
