### 2019-08-24 imr-framework/virtual-scanner#23 @vsoch
hey @sairamgeethanath just wanted to see how you are doing here? I had forgotten about this issue, and went to continue my review, and then (again) realized I didn't want to install all the dependencies on my host (and remembered). Do you need help?
