### 2019-08-24 222f614
adding Dockerfile for container usage
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
