### 2019-08-24 imr-framework/virtual-scanner#23 @mathieuboudreau
@vsoch fyi I opened an issue (#25) in my own review about this too. I recommended at least adding instructions to make a virtual environment using `venv` or `conda` (if available), and the authors updated the readme to recommend using `venv`. I have another issue opened related to installation/setup in issue #29
