### 2019-08-24 imr-framework/virtual-scanner#23 @mathieuboudreau
Actually I see that they removed the reference to `venv` and not just say to create a virtual environment using the tool of your choice.
