### 2019-08-24 imr-framework/virtual-scanner#23 @vsoch
It's okay, I went ahead and added a basic container with instructions here - https://github.com/imr-framework/virtual-scanner/pull/35 do you want to test it out and make sure everything runs the same as it should locally? I can update the PR if there is a bug.
