### 2019-08-24 imr-framework/virtual-scanner#36 @vsoch
JoSS Review: Using Virtual Scanner
This is for https://github.com/openjournals/joss-reviews/issues/1637 and note that I'm testing with the container from #35. These are some issues related to using the software.

## Login
As a user, I'm tempted to just select standard or advanced and click "Start Scan" but nothing happens. I want to suggest any of the following:

 - Don't require an email to use it (this seems like an unnecessary bit of personal information for a user to provide to a development server).
 - If the email must be required (why?) disable the button until it's entered.
 - if the email isn't entered and I click start scan, I should see the box highlighted in red and/or a notification that I need to enter my email before continuing.

We want to avoid the current setup where I can click the button and get no response from the interface about anything being wrong.

## Subject Entering
When I first log in, it's very intuitive to enter a subject (I made a bunch of fields up) on the Register screen. I then proceeded to the Acquisition. However, when I selected the sequence and then clicked "Run Scan" it jumps me back to the register page and tells me I need to register a subject. I "think" what it needs me to do is select "Numerical" from the dropdown (register a phantom?) and then continue, but it's not obvious from this usage. I'd like to suggest that if I need to register a phantom (or numerical) it takes me back to that page, with that specific error message, and the Numerical tab selected instead of Subject.

## Run Scan

Once I entered a numerical subject, I clicked Run Scan and there was an error (and here is the traceback - an index out of bounds):

```
Traceback (most recent call last):
  File "/opt/conda/lib/python3.7/site-packages/flask/app.py", line 2309, in __call__
    return self.wsgi_app(environ, start_response)
  File "/opt/conda/lib/python3.7/site-packages/flask/app.py", line 2295, in wsgi_app
    response = self.handle_exception(e)
  File "/opt/conda/lib/python3.7/site-packages/flask/app.py", line 1741, in handle_exception
    reraise(exc_type, exc_value, tb)
  File "/opt/conda/lib/python3.7/site-packages/flask/_compat.py", line 35, in reraise
    raise value
  File "/opt/conda/lib/python3.7/site-packages/flask/app.py", line 2292, in wsgi_app
    response = self.full_dispatch_request()
  File "/opt/conda/lib/python3.7/site-packages/flask/app.py", line 1815, in full_dispatch_request
    rv = self.handle_user_exception(e)
  File "/opt/conda/lib/python3.7/site-packages/flask/app.py", line 1718, in handle_user_exception
    reraise(exc_type, exc_value, tb)
  File "/opt/conda/lib/python3.7/site-packages/flask/_compat.py", line 35, in reraise
    raise value
  File "/opt/conda/lib/python3.7/site-packages/flask/app.py", line 1813, in full_dispatch_request
    rv = self.dispatch_request()
  File "/opt/conda/lib/python3.7/site-packages/flask/app.py", line 1799, in dispatch_request
    return self.view_functions[rule.endpoint](**req.view_args)
  File "/opt/conda/lib/python3.7/site-packages/virtual_scanner-1.0.21-py3.7.egg/virtualscanner/coms/coms_ui/coms_server_flask.py", line 339, in worker
    progress = bsim.run_blochsim(seqinfo=payload, phtinfo=rows[0][0],
IndexError: list index out of range
172.17.0.1 - - [24/Aug/2019 19:56:21] "GET /receiver?__debugger__=yes&cmd=resource&f=style.css HTTP/1.1" 200 -
```
Likely this is because some input parameter was incorrect / missing, so my suggestion is to do validation of all parameters first, and then return to the form if anything is invalid. In fact, this shouldn't happen only on the Run Scan page, but also on the Register page - I shouldn't be allowed to register with missing / incorrect data, or to click Run Scan without having a validation first.

I'll continue my review after these are updated - feel free to merge #35 so you can start with the updated base (and make adding these next changes easier, without merge conflicts, etc.)
