### 2019-08-24 imr-framework/virtual-scanner#23 @sairamgeethanath
@sravan did do a docker image and tried it yesterday which was buggy, we
will fix it on Monday.

We also have modified the readme to suit multiple install options.

Will respond in detail on Monday.
