### 2019-08-24 imr-framework/virtual-scanner#23 @vsoch
@sairamgeethanath I did a pull request with a fully working Dockerfile, see #35. I've been able to continue my review in #36. #35 is good to go, if you want to merge and then work on changes for #36. Thanks!
