### 2019-08-24 imr-framework/virtual-scanner#23 @sairamgeethanath
That would be great, will check it out and respond soon.
