### 2019-08-29 cb80226
Issue#36. Add default email to log-in page + Clarify Subject: Numerical
