### 2019-08-29 imr-framework/virtual-scanner#36 @tonggehua
@vsoch they should be resolved now with cb802265e0e337eec3deff3e7c2ec3faa0b4685e
