### 2019-08-29 imr-framework/virtual-scanner#36 @vsoch
Thanks! I'm pretty busy atm so I'll need to pick this up later. It's "kept as new" in my inbox :)
