### 2019-09-09 imr-framework/virtual-scanner#36 @sairamgeethanath
Given that we are unable to dockerize the code and reproduce errors from docker in the local environment, should we do the docker for the next version of the release? @tonggehua @sravan953 @Mmj94 @vsoch @mathieuboudreau @qianenlin
