### 2019-09-09 imr-framework/virtual-scanner#29 @sairamgeethanath
@mathieuboudreau Can we close #29 separately now as it shows up still as open?
