### 2019-09-09 imr-framework/virtual-scanner#36 @mathieuboudreau
I'm satisfied with at least having a virtual environment setup, I'm not completely attached to having a Docker one in this release.
