### 2019-09-09 imr-framework/virtual-scanner#36 @vsoch
I didn't have any issues with the Docker recipe I produced... Dockerization is a requirement on my end.
