### 2019-09-09 imr-framework/virtual-scanner#36 @vsoch
Virtual environments are one step up from not providing anything, but virtual environments != reproduciblity. I [mentioned this requirement](https://github.com/openjournals/joss-reviews/issues/1590#issuecomment-515681126) when I agreed to review, so I hope it is not coming as a surprise. I can help in whatever way is necessary, perhaps you can start by sharing what is different about the container and running locally?
