### 2019-09-09 imr-framework/virtual-scanner#36 @tonggehua
We are not sure what is different between the container and running locally - not enough to resolve the bug. The crash mentioned earlier in https://github.com/imr-framework/virtual-scanner/issues/36#issue-484857122 and also a similar crash on Analyze page both happen only when we use Docker, not when we are running within the local virtual environment, with Pycharm. We've tried to debug with Docker and Pycharm but ran into problems. In addition, the supported version of Docker doesn't run on Windows 10 Home which is what two of our lab mates use, and likely many of our potential users. We are open to suggestions on how to debug :)
