### 2019-09-09 imr-framework/virtual-scanner#36 @vsoch
Windows still has a lot of issues with using Docker - do you have a linux machine to test it on? Or even a Mac? Or can you grab a cloud instance and run there, exposing the port to your local machine?

The target audience is usually not a Windows user, as containers and windows are a relatively (still) buggy thing, but that of course is my opinion. I would imagine someone on linux wanting to run it (or more likely) wanting to deploy it for a lab. [Most deployment servers are linux](https://www.zdnet.com/article/linux-foundation-finds-enterprise-linux-growing-at-windows-expense/).

As a sidenote, if your entire lab is truly running windows (which makes sense, I came from a neuropsych lab and everyone was using Windows and Matlab) I would strongly consider, even for a small personal machine, having linux (ubuntu flavor or similar even) to use for programming and development. My learning increased exponentially in graduate school when my (previous) windows machine crashed, and I really was forced to work on a command line, compile things, etc.
