### 2019-09-09 6a19c16
Fix Dockerfile
- Pip install will build from local source code
- Use ADD instead of COPY
- Register, acquire and analyze tested successfully
- Build successful on Mac
