+++ Dockerfile
@@ -1,12 +1,16 @@
-FROM continuumio/miniconda3
-# docker build -t virtualscanner .
-RUN apt-get update && \
-    apt-get install -y libglib2.0-0 \
-                       libsm6 libxext6 libxrender-dev && \
-    mkdir -p /code
-ADD . /code
-RUN pip install -r /code/requirements.txt
-WORKDIR /code
-RUN python setup.py install
+# Use an official Python runtime as a parent image
+FROM python:3.6.3
+
+# Add source code to /virtual-scanner
+ADD . /virtual-scanner
+
+# Set working directory to /virtual-scanner
+WORKDIR /virtual-scanner
+
+# Install requirements
+RUN pip install --trusted-host pypi.python.org -r requirements.txt
+
+# Expose Virtual Scanner's port 5000
 EXPOSE 5000
-ENTRYPOINT ["/opt/conda/bin/virtualscanner"]
+
+CMD ["python", "virtualscanner/coms/coms_ui/coms_server_flask.py"]
