### 2019-09-09 5075225
Add OS-specific instructions for Docker
