### 2019-09-10 imr-framework/virtual-scanner#36 @tonggehua
Crash errors and Docker build problems were fixed with 6a19c16d4496dd036e94f077c6584c0ad735e700, with a new Dockerfile update. Right now, Windows users should only use the published image (there is an OSError when trying to run a locally built image) and Mac users can either use the published image or build locally (instructions in README.md have been updated). We currently do not have a linux setup but would love to look into it if needed. Please review. Again, thanks for the suggestions!
