### 2019-09-10 imr-framework/virtual-scanner#36 @vsoch
Can you please explain what you mean by not having a Linux set up? Containers run native we on Linux, much more easily than Mac or Windows.  Original recipe that I developed ran cleanly on Ubuntu.
