### 2019-09-10 imr-framework/virtual-scanner#36 @vsoch
If you just need a test of the new recipe on Linux I’d be happy to do this tomorrow :)
