### 2019-09-10 imr-framework/virtual-scanner#36 @vsoch
oy vey, my work wireless is still out (and I'm teathered to my phone) so I'll need to wait to test this for when I go back to my apartment later this afternoon. Even cloning a git repo is substantially slow / takes a lot of bandwidth!
