### 2019-09-10 imr-framework/virtual-scanner#36 @tonggehua
@vsoch Thanks so much for the response! By "not having a Linux setup", I mean that we don't a Linux system installed and in use for research on lab/personal computers.
