### 2019-09-10 imr-framework/virtual-scanner#29 @tonggehua
@mathieuboudreau Any updates on the README.md review? The latest update contains instructions for using the published Docker image as well. That makes three ways of getting started.
