### 2019-09-10 imr-framework/virtual-scanner#29 @mathieuboudreau
I had booked some time this afternoon to catch up on that and other parts of your review, I'll let you know ASAP!
