### 2019-09-10 imr-framework/virtual-scanner#36 @vsoch
ah okay! I thought you meant that the container wouldn't work for linux (it should!) I'll let you know when I'm able to test.
