### 2019-09-10 imr-framework/virtual-scanner#29 @mathieuboudreau
@tonggehua FYI, if I follow the instructions under "Quick Start" and follow the typical UX path (email, subject, select sequence, then run exam), I get the following error:

I'm on Mac 10.14.6 and Safari 12.1.2. Could you check to see if these instructions work for you?
