### 2019-09-10 imr-framework/virtual-scanner#29 @mathieuboudreau
I just tried the "Pro start" again. A few notes:

1 - As mentionned previously, the line "Run `coms_server_flask.py` to run your changes." isn't clear enough, since the file is hidden under several folders. I recommend you change it to "Run `python virtualscanner/coms/coms_ui/coms_server_flask.p` ..." to make the life of your new users a bit easier.

2- After running it, I noticed that the user interface & functionality is different from the "Quick start" mode, so make sure to try and keep your pypi releases up to date.

3 - I can select a pulse sequence, but now "Run exam" is unclickable for me, so I can't actually run the example. Is this just a problem on my end?
