### 2019-09-10 imr-framework/virtual-scanner#29 @mathieuboudreau
If it works for you and doesn't for me, I'll take a video to show you what I'm experiencing
