### 2019-09-10 imr-framework/virtual-scanner#29 @mathieuboudreau
Tried the Docker instructions too, and I hit the same wall of not being able to click "Run exam""
