### 2019-09-10 imr-framework/virtual-scanner#29 @mathieuboudreau
Ahh, I just tried on Chrome and it works fine - stupid Safari. Sorry about that. The only point that doesn't work for me is the first one (Quick Start).
