### 2019-09-11 imr-framework/virtual-scanner#36 @vsoch
heyo! I didn't get to test it today, but I made sure to find a spot with a good connection, clone the repo, and build the container - so I should be able to complete my test and review tomorrow morning.
