### 2019-09-11 imr-framework/virtual-scanner#36 @vsoch
okay testing! The start up went much more smoothly this time - I could click a mode and proceed without worrying about email. The subject was then largely set up and I just needed to continue to analyze, and then the analyze (albeit a bit slow, I know you'll work on it) did produce an Axial image. I tested both GE and SE, and also the Analyze tab with the sample data. I'm not totally clear on what this is useful for, but I'm not actively working in this area so I don't expect to have this.

For the future - it could be that the slowness might be helped via using some sort of multiprocessing, or conversion to a Singularity container that can better leverage resources on the host (e.g., gpu libraries). My last suggestion is to put somewhere a note about expected processing times, because they are upwards of minutes, at least testing in the Docker container.

Otherwise, great job on the fixes here! I'm looking forward to seeing how this is further developed.
