### 2019-09-11 imr-framework/virtual-scanner#29 @tonggehua
@mathieuboudreau what version of pip-installed Virtual Scanner are you using?
