### 2019-09-11 imr-framework/virtual-scanner#29 @mathieuboudreau
The latest one (1.0.22). It's dated from August 15th on pypi.
