### 2019-09-13 imr-framework/virtual-scanner#29 @tonggehua
@sravan953 Maybe this is a quick fix?
