### 2019-09-13 88131ff
PyPI compatible README.md
