### 2019-09-14 imr-framework/virtual-scanner#25 @mathieuboudreau
FYI, the issue is still not resolved for me. Trying the latest `pip` version of your software (1.0.24), I can't even access the server now, I just get this error on my terminal when running `virtualscanner`:

```
Using TensorFlow backend.
Traceback (most recent call last):
  File "/Users/mathieuboudreau/anaconda3/envs/vs_pypi/bin/virtualscanner", line 10, in <module>
    sys.exit(launch_virtualscanner())
  File "/Users/mathieuboudreau/anaconda3/envs/vs_pypi/lib/python3.6/site-packages/virtualscanner/coms/coms_ui/coms_server_flask.py", line 526, in launch_virtualscanner
    __modify_search_path()
  File "/Users/mathieuboudreau/anaconda3/envs/vs_pypi/lib/python3.6/site-packages/virtualscanner/coms/coms_ui/coms_server_flask.py", line 9, in __modify_search_path
    SEARCH_PATH = script_path[:script_path.index('virtual-scanner') + len('virtual-scanner') + 1]
ValueError: substring not found
```

On a side note: it may be a good idea for you folks to setup a continuous testing suite, using something like Travis-CI (free). Several commits made during this JOSS review introduced new bugs that were later found by manual testing from external users (like me), so for better software stability it would be useful if you were alerted immediately before users experience these bugs. Another thing that can improve stability is not pushing directly to master, but instead pushing to branches and only merging to master after the tests pass (and ideally, a code review too). JOSS even recommends this in their review guideline (though it's not required, I'm just pointing out how beneficial for you it would be to spend a little bit of time setting this up, for long term benefits).
