### 2019-09-14 imr-framework/virtual-scanner#25 @mathieuboudreau
Dug in a bit deeper, this recent commit works if I run `virtualscanner` after I `git clone` the repo and install it using `pip install -e .`, just not if I install it through `pip install virtual-scanner`. Maybe has to do with the fact that when installed using `pip install virtual-scanner`, it's installed in a folder called `virtualscanner` and not `virtual-scanner` in the venv folder (i.e. only your module/subfolder [virtualscanner](https://github.com/imr-framework/virtual-scanner/tree/master/virtualscanner) is copied with this installation, vs. the `pip install -e .` installation which links to the cloned folder ?
