### 2019-09-16 imr-framework/virtual-scanner#29 @mathieuboudreau
Odd, I was sure I sent a message on Friday about this. Did I post it in the wrong issue?

No, that fix did not resolve my issue completely. Now, when installing using `pip`, the command `virtualscanner` crashes before the server is even booted up, see error message below:

```
Using TensorFlow backend.
Traceback (most recent call last):
  File "/Users/mathieuboudreau/anaconda3/envs/vs_pypi/bin/virtualscanner", line 10, in <module>
    sys.exit(launch_virtualscanner())
  File "/Users/mathieuboudreau/anaconda3/envs/vs_pypi/lib/python3.6/site-packages/virtualscanner/coms/coms_ui/coms_server_flask.py", line 526, in launch_virtualscanner
    __modify_search_path()
  File "/Users/mathieuboudreau/anaconda3/envs/vs_pypi/lib/python3.6/site-packages/virtualscanner/coms/coms_ui/coms_server_flask.py", line 9, in __modify_search_path
    SEARCH_PATH = script_path[:script_path.index('virtual-scanner') + len('virtual-scanner') + 1]
ValueError: substring not found
```

However, if I clone the repo and install it using `pip install -e .`, it works fine. I think the reason being that you're seaching for the path `virtual-scanner` which is the **repository name**, and not the **package name** (`virtualscanner`, the folder within your repository). When installing using pypi, only the **package folder** is installed into your computer, so it won't find `virtual-scanner` in the users path.

p.s. This circular workflow (user reports bug -> dev fixes bug -> user finds another bug due to previous bug fix) shows why it might be useful to test your software using some kind of free continuous integration suite like Travis-CI (which would have flagged it during the "dev fixes bug" process in an email alert, instead of waiting until the next user encounters it). JOSS recommends it in its reviewing guidelines, but doesn't make it a requirement. Another helpful step would be to not push directly to master your changes, but create changes in branches and merge those into master using pull requests & code review/continuous integration review.
