### 2019-09-16 imr-framework/virtual-scanner#29 @sravan953
@mathieuboudreau Sorry for the inconvenience, and thank you for your inputs about automated testing/CI. We are discussing and looking at ways to incorporate it in our workflow. For now, we should have fixed the startup issues whether it is a `pip install -e .`, `pip install virtual-scanner` or running in IDE. Eagerly waiting to hear back.
