### 2019-09-16 imr-framework/virtual-scanner#29 @mathieuboudreau
As I said above, `pip install -e .` works for me, `pip install virtual-scanner` does not. I explain why I suspect the latter fails in my above message.
