### 2019-09-16 imr-framework/virtual-scanner#29 @sravan953
Yes, it took some digging around to figure it out, but turns out the upload to PyPI was not packaging a crucial asset file along with the source code, and hence the crash. Should be fixed now!
