### 2019-09-16 12da3c2
Include subject.db for pip install
