### 2019-09-16 imr-framework/virtual-scanner#29 @sairamgeethanath
Path issues again  :(

On Mon, Sep 16, 2019, 12:34 Mathieu Boudreau <notifications@github.com>
wrote:

> Getting close!
>
> The server runs now with the new pypi release. I can Login, register the
> subject, load a sequence. However when I run the exam, it just freezes
> there and doesn't output the image. Here is my log (similar error as above
> occurs):
>
> Using TensorFlow backend.
>  * Serving Flask app "virtualscanner.coms.coms_ui.coms_server_flask" (lazy loading)
>  * Environment: production
>    WARNING: Do not use the development server in a production environment.
>    Use a production WSGI server instead.
>  * Debug mode: on
>  * Running on http://0.0.0.0:5000/ (Press CTRL+C to quit)
>  * Restarting with stat
> Using TensorFlow backend.
>  * Debugger is active!
>  * Debugger PIN: 152-984-311
> 127.0.0.1 - - [16/Sep/2019 13:59:43] "GET / HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:43] "GET /static/main_style.css HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:43] "GET /static/packages/jquery-3.3.1.min.js HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:43] "GET /static/log_in_style.css HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:43] "GET /static/virtual_scanner_logo_v1.png HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:43] "GET /static/log_in_func.js HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:43] "GET /static/logo-blue-white-onlycrown.png HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:45] "POST / HTTP/1.1" 302 -
> 127.0.0.1 - - [16/Sep/2019 13:59:45] "GET /register HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:45] "GET /static/packages/Bootstrap/jquery-3.3.1.slim.min.js HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:45] "GET /static/packages/Bootstrap/popper.min.js HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:45] "GET /static/packages/Bootstrap/bootstrap.min.css HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:45] "GET /static/packages/Bootstrap/bootstrap.min.js HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:45] "GET /static/packages/jQuery-ui/jquery-ui.css HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:45] "GET /static/packages/jQuery-ui/jquery-ui.js HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:45] "GET /static/register_style.css HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:45] "GET /static/register_func.js HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:45] "GET /static/phantomAutofill.js HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:45] "GET /static/log_out_func.js HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:45] "GET /static/phantom_pics/Numerical.png HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:46] "GET /acquire HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:47] "GET /static/acquire_func.js HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:47] "GET /static/acquire_style.css HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:51] "POST /receiver HTTP/1.1" 302 -
> 127.0.0.1 - - [16/Sep/2019 13:59:51] "GET /register HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:53] "GET /acquire HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 13:59:59] "POST /receiver HTTP/1.1" 302 -
> 127.0.0.1 - - [16/Sep/2019 13:59:59] "GET /register HTTP/1.1" 200 -
> {'formName': 'reg', 'subjecttype': 'Numerical', 'patid': '2321', 'name': 'Numerical', 'age': '00', 'dob': '9/16/2019', 'gender': 'other', 'weight': '3', 'weight-unit': 'kg', 'height': '20', 'height-unit': 'cm', 'inches': '', 'orient': 'HFS', 'anatomy': 'brain'}
> 0
> 127.0.0.1 - - [16/Sep/2019 14:00:01] "POST /receiver HTTP/1.1" 302 -
> 127.0.0.1 - - [16/Sep/2019 14:00:01] "GET /register_success HTTP/1.1" 200 -
> 127.0.0.1 - - [16/Sep/2019 14:00:02] "GET /acquire HTTP/1.1" 200 -
> [('Numerical', 2321, 'Numerical', 0, '9/16/2019', 'other', 3.0, 20.0, 'HFS', 'brain')]
> {'formName': 'acq', 'selectedSeq': 'GRE', 'TR': '200', 'TE': '10', 'FA': '90', 'sl-orient': 'axial', 'thck': '16', 'slicenum': '1', 'freq': 'x', 'ph': 'y', 'Nx': '15', 'Ny': '15', 'FOVx': '240', 'FOVy': '240'}
> Traceback (most recent call last):
>   File "/Users/mathieuboudreau/anaconda3/envs/vs10/lib/python3.6/site-packages/virtualscanner/server/simulation/bloch/pulseq_bloch_simulator.py", line 18, in <module>
>     SEARCH_PATH = script_path[:script_path.index('virtual-scanner') + len('virtual-scanner') + 1]
> ValueError: substring not found
>
