### 2019-09-17 imr-framework/virtual-scanner#36 @sairamgeethanath
@vsoch Could you please let us know if you are okay for us to close this issue?
