### 2019-09-17 imr-framework/virtual-scanner#36 @vsoch
Yes we're good! Although I haven't finished the review yet - I'm technically taking vacation days but I'll see if I can squeeze it in soon.
