### 2019-09-17 imr-framework/virtual-scanner#38 @vsoch
JoSS Review: Testing
This is for https://github.com/openjournals/joss-reviews/issues/1637. The last major requirement for this repository is to add proper testing, either via Travis, CircleCI, or another service of your choosing. Especially given the errors that came up (and a file missing from the release) it would be best to do a minimal build, and set of tests for the core software. There are a lot of really good starter example templates, and it's fairly easy to connect and setup.
