### 2019-09-19 #1637 @arokem
Where do things stand here?

@vsoch: are you able to check any of the boxes on your list?

@mathieuboudreau  and @nstikov: from your end, I see that the only remaining points are installation and version. Not worrying about version for the moment, what still needs to be done for the installation check box? Is there an associated issue on the project repo?
