### 2019-09-19 #1637 @vsoch
I'm still reviewing - I'll check them when I'm finished with the issues above! For example, I'm waiting to hear back about testing (issue 38 linked above). I definitely am not dragging feet or trying to be slow, I'm bringing up issues and the authors are addressing them, and I'll update the final review here when that is done.
