### 2019-09-19 #1637 @mathieuboudreau
There is an issue on my end, but the authors closed it without it fully being resolved. It's https://github.com/imr-framework/virtual-scanner/issues/29

Their "Quick Start" installation instructions are still broken as of the last time I tried it out.
