### 2019-09-19 #1637 @arokem
Thanks for the feedback!

> Their "Quick Start" installation instructions are still broken as of the last time I tried it out.

@imr-framework : I agree with @mathieuboudreau. Looks like that issue is still not resolved (despite an impressive amount of work on the part of the reviewer and the authors to resolve this!). Could you please work to resolve this issue? It looks like it's getting closer...
