### 2019-09-19 #1637 @mathieuboudreau
If it does get too cumbersome to fix this issue, I'm fine with them just omitting that section from their README, as there are two other methods of installations describe on there that do work. I just don't want there to be instructions that don't work listed (particular the "easiest" one, which might discourage new users).
