### 2019-09-19 #1637 @vsoch
+1! And this is a really good (real life example) for why I pushed hard for containers. Virtual environments are better than nothing, but they aren't reproducible in the same way a container is.
