+++ .github/workflows/pythonpackage.yml
@@ -5,7 +5,7 @@ on: [push]
 jobs:
   build:
 
-    runs-on: [ubuntu-latest, windows-latest, macOS-latest]
+    runs-on: windows-latest
     strategy:
       max-parallel: 2
       matrix:
