### 2019-09-27 imr-framework/virtual-scanner#38 @tonggehua
@vsoch We are working on this with Github Actions. What level of testing would you suggest? We have unit tests for the back-end code and plan to include those as part of the CI workflow. We have encountered errors from the GUI frequently though - should there be automated GUI tests too?
