### 2019-09-27 imr-framework/virtual-scanner#38 @vsoch
Front end testing is hard, so I would do unit tests for what you are able to do, and then test build / install and deploy of the interface. With selenium (or similar) you can do basic page navigation and just confirm that the response is successful (e.g., 200).
