### 2019-10-02 b335744
Remove pypulseq2
