+++ virtualscanner/pypulseq2/calc_ramp.py
@@ -1,285 +1,0 @@
-import numpy as np
-
-from pypulseq.opts import Opts
-
-
-def calc_ramp(k0, k_end, system=Opts(), max_points=500, max_grad=np.zeros(0), max_slew=np.zeros(0)):
-    def inside_limits(grad, slew):
-        if mode == 0:
-            grad2 = np.sum(np.square(grad), axis=1)
-            slew2 = np.sum(np.square(slew), axis=1)
-            ok = np.all(np.max(grad2) <= np.square(max_grad)) and np.all(np.max(slew2) <= np.square(max_slew))
-        else:
-            ok = (np.sum(np.max(np.abs(grad), axis=1) <= max_grad) == 3) and (
-                    np.sum(np.max(np.abs(slew), axis=1) <= max_slew) == 3)
-
-        return ok
-
-    def __joinleft0(k0, k_end, G0, G_end, use_points):
-        if use_points == 0:
-            G = np.stack((G0, (k_end - k0) / grad_raster, G_end)).T
-            S = (G[:, 1:] - G[:, :-1]) / grad_raster
-
-            k_out_left = np.zeros((3, 0))
-            success = inside_limits(G, S)
-
-            return success, k_out_left
-
-        dk = (k_end - k0) / (use_points + 1)
-        kopt = k0 + dk
-        Gopt = (kopt - k0) / grad_raster
-        Sopt = (Gopt - G0) / grad_raster
-
-        okGopt = np.sum(np.square(Gopt)) <= np.square(max_grad)
-        okSopt = np.sum(np.square(Sopt)) <= np.square(max_slew)
-
-        if okGopt and okSopt:
-            k_left = kopt
-        else:
-            a = np.multiply(max_grad, grad_raster)
-            b = np.multiply(max_slew, grad_raster ** 2)
-
-            dkprol = G0 * grad_raster
-            dkconn = dk - dkprol
-
-            ksl = k0 + dkprol + dkconn / np.linalg.norm(dkconn) * b
-            Gsl = (ksl - k0) / grad_raster
-            okGsl = np.sum(np.square(Gsl)) <= np.square(max_grad)
-
-            kgl = k0 + np.multiply(dk / np.linalg.norm(dk), a)
-            Ggl = (kgl - k0) / grad_raster
-            Sgl = (Ggl - G0) / grad_raster
-            okSgl = np.sum(np.square(Sgl)) <= np.square(max_slew)
-
-            if okGsl:
-                k_left = ksl
-            elif okSgl:
-                k_left = kgl
-            else:
-                c = np.linalg.norm(dkprol)
-                c1 = np.divide(np.square(a) - np.square(b) + np.square(c), (2 * c))
-                h = np.sqrt(np.square(a) - np.square(c1))
-                kglsl = k0 + np.multiply(c1, np.divide(dkprol, np.linalg.norm(dkprol)))
-                projondkprol = (kgl * dkprol.T) * (dkprol / np.linalg.norm(dkprol))
-                hdirection = kgl - projondkprol
-                kglsl = kglsl + h * hdirection / np.linalg.norm(hdirection)
-                k_left = kglsl
-
-        success, k = __joinright0(k_left, k_end, (k_left - k0) / grad_raster, G_end, use_points - 1)
-        if len(k) != 0:
-            if len(k.shape) == 1:
-                k = k.reshape((len(k), 1))
-            if len(k_left.shape) == 1:
-                k_left = k_left.reshape((len(k_left), 1))
-            k_out_left = np.hstack((k_left, k))
-        else:
-            k_out_left = k_left
-
-        return success, k_out_left
-
-    def __joinleft1(k0, k_end, G0, G_end, use_points):
[…]
