+++ virtualscanner/pypulseq2/Sequence/read_seq.py
@@ -1,209 +1,0 @@
-import numpy as np
-
-from pypulseq.event_lib import EventLibrary
-
-
-def read(self, path, **kwargs):
-    """
-    Reads .seq file from path, and constructs a Sequence object from the file.
-
-    Parameters
-    ----------
-    path : str
-        Path of .seq file to be read.
-    """
-
-    detect_rf_use = True if 'detect_rf_use' in kwargs else False
-
-    input_file = open(path, 'r')
-    self.shape_library = EventLibrary()
-    self.rf_library = EventLibrary()
-    self.grad_library = EventLibrary()
-    self.adc_library = EventLibrary()
-    self.delay_library = EventLibrary()
-    self.block_events = {}
-    self.rf_raster_time = self.system.rf_raster_time
-    self.grad_raster_time = self.system.grad_raster_time
-    self.definitions = {}
-
-    while True:
-        section = skip_comments(input_file)
-        if section == -1:
-            break
-
-        if section == '[DEFINITIONS]':
-            self.definitions = read_definitions(input_file)
-        elif section == '[VERSION]':
-            version_major, version_minor, version_revision = read_version(input_file)
-            if version_major != self.version_major or version_minor != self.version_minor or version_revision != self.version_revision:
-                raise Exception(
-                    f'Unsupported version: {version_major, version_minor, version_revision}. '
-                    f'Expected: {self.version_major, self.version_minor, self.version_revision}')
-        elif section == '[BLOCKS]':
-            self.block_events = read_blocks(input_file)
-        elif section == '[rf]':
-            self.rf_library = read_events(input_file, [1, 1, 1, 1e-6, 1, 1])
-        elif section == '[GRAD]':
-            self.grad_library = read_events(input_file, [1, 1, 1e-6], 'g', self.grad_library)
-        elif section == '[TRAP]':
-            self.grad_library = read_events(input_file, [1, 1e-6, 1e-6, 1e-6, 1e-6], 't', self.grad_library)
-        elif section == '[ADC]':
-            self.adc_library = read_events(input_file, [1, 1e-9, 1e-6, 1, 1])
-        elif section == '[DELAYS]':
-            self.delay_library = read_events(input_file, 1e-6)
-        elif section == '[SHAPES]':
-            self.shape_library = read_shapes(input_file)
-        else:
-            raise ValueError(f'Unknown section code: {section}')
-
-    if detect_rf_use:
-        for k in self.rf_library.keys():
-            lib_data = self.rf_library.data[k]
-            rf = self.rf_from_lib_data(lib_data)
-            flip_deg = abs(np.sum(rf.signal)) * rf.t[0] * 360
-            if len(lib_data) < 9:
-                if flip_deg <= 90:
-                    lib_data[8] = 0
-                else:
-                    lib_data[8] = 2
-                self.rf_library.data[k] = lib_data
-
-
-def read_definitions(input_file):
-    definitions = dict()
-    line = strip_line(input_file)
-    while line != '' and line[0] != '#':
-        tok = line.split(' ')
-        if not any([x.isalpha() for x in tok[1:]]):
-            definitions[tok[0]] = np.array(tok[1:], dtype=float)
-        else:
-            definitions[tok[0]] = tok[1:]
-        line = strip_line(input_file)
-
-    return definitions
-
-
-def read_version(input_file):
-    line = strip_line(input_file)
-    while line != '' and line[0] != '#':
-        tok = line.split(' ')
-        if tok[0] == 'major':
-            major = np.array(tok[1:], dtype=float)
-        elif tok[0] == 'minor':
-            minor = np.array(tok[1:], dtype=float)
-        elif tok[0] == 'revision':
-            revision = np.array(tok[1:], dtype=float)
-        line = strip_line(input_file)
-
[…]

[…]
