### 2019-10-15 443e25b
Auto stash before rebase of "origin/master"
Included the unit test class to include in CI
