### 2019-10-15 9259354
Modify utest discovery file name criterion
