+++ .github/workflows/pythonpackage.yml
@@ -29,6 +29,7 @@ jobs:
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test (WIP)
-      run: python -m unittest
+      run: python -m unittest discover -p "utest*.py"
+      #run: python -m unittest
       #pip install pytest
       #pytest
