### 2019-10-16 a14550b
Add utests for bloch
