### 2019-10-16 f98de30
Add mac and ubuntu to system matrix
