+++ .github/workflows/pythonpackage.yml
@@ -10,14 +10,13 @@ jobs:
       max-parallel: 2
       matrix:
         os: [macOS-lastest, windows-lastest, ubuntu-latest]
-        python-version: [3.6]
    
     steps:
     - uses: actions/checkout@v1
     - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v1
       with:
-        python-version: ${{ matrix.python-version }}
+        python-version: 3.6
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
