### 2019-10-16 3518ea8
Change mapping utest to run on multiple os
