+++ .github/workflows/pythonpackage.yml
@@ -13,7 +13,7 @@ jobs:
    
     steps:
     - uses: actions/checkout@v1
-    - name: Set up Python ${{ matrix.python-version }}
+    - name: Set up Python 3.6
       uses: actions/setup-python@v1
       with:
         python-version: 3.6
