### 2019-10-16 imr-framework/virtual-scanner#29 @tonggehua
@mathieuboudreau  Please try the pip install with the new 1.0.26 version. There shouldn't be a problem with seeing the simulated images now. We did find a problem with using the "virtualscanner" command on Windows (it still works on macOS) but will add the command that lets the user run coms_server_flask.py locally after a pip install.
