### 2019-10-16 68d9ca9
Include Windows Quickstart instructions
