### 2019-10-17 imr-framework/virtual-scanner#29 @mathieuboudreau
Excellent!! This installation procedure works perfectly now on my system.
