### 2019-10-17 #1637 @mathieuboudreau
@imr-framework has resolved the outstanding issues I had with the installation instructions, and I've now completed all of the items on my checklist. Thanks a lot @imr-framework for making my suggested fixes/desired features along the way!
