### 2019-10-17 #1637 @arokem
Thanks @mathieuboudreau ! @vsoch : I see that none of your boxes are checked at this point. Have you had a chance to take a look? Do you still have any concerns or unadressed issues?
