### 2019-10-17 #1637 @vsoch
oups, sorry about that! I think I must have checked off the boxes in a browser tab many weeks ago and forgot to press save :)

I've updated my checkboxes, the only missing item is automated tests, represented in this issue https://github.com/imr-framework/virtual-scanner/issues/38
