### 2019-10-18 e2a7708
Try conda install matplotlib
