+++ .github/workflows/pythonpackage.yml
@@ -21,6 +21,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install -r requirements.txt
+        conda install matplotlib
     - name: Lint with flake8
       run: |
         pip install flake8
