+++ .github/workflows/pythonpackage.yml
@@ -9,7 +9,7 @@ jobs:
     strategy:
       max-parallel: 2
       matrix:
-        os: [windows-latest, macOS-latest]
+        os: [windows-latest]
    
     steps:
     - uses: actions/checkout@v1
@@ -21,7 +21,6 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install -r requirements.txt
-        conda install matplotlib
     - name: Lint with flake8
       run: |
         pip install flake8
