### 2019-10-29 #1637 @tonggehua
@arokem Sorry for the late reply. My research group is focusing on a conference abstract deadline (Nov. 6) and we plan to address the automated testing issue immediately after it. Most back-end tests are done, but setting up proper front-end testing (with Selenium) will take more time.
