+++ .github/workflows/pythonpackage.yml
@@ -9,7 +9,8 @@ jobs:
     strategy:
       max-parallel: 2
       matrix:
-        os: [windows-latest]
+        os: [macOS-latest]
+        #os: [windows-latest]
    
     steps:
     - uses: actions/checkout@v1
