+++ .github/workflows/pythonpackage.yml
@@ -17,7 +17,7 @@ jobs:
     - name: Set up Python 3.6
       uses: actions/setup-python@v1
       with:
-        python-version: 3.6
+        python-version: 3.6.5
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
