+++ .github/workflows/pythonpackage.yml
@@ -17,7 +17,7 @@ jobs:
     - name: Set up Python 3.6
       uses: actions/setup-python@v1
       with:
-        python-version: 3.5.7
+        python-version: 3.6
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
@@ -30,7 +30,7 @@ jobs:
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test (WIP)
-      run: python -m unittest discover -p "utest*.py"
+      run: python -m unittest discover -p "utest_pulseq*.py"
       #run: python -m unittest
       #pip install pytest
       #pytest
