### 2019-11-13 be3e046
Update matplotlib import for macOS
