+++ .github/workflows/pythonpackage.yml
@@ -9,8 +9,8 @@ jobs:
     strategy:
       max-parallel: 2
       matrix:
-        os: [macOS-latest]
-        #os: [windows-latest]
+        os: [windows-latest]
+        #os: [macOS-latest]
    
     steps:
     - uses: actions/checkout@v1
