### 2019-11-13 50a25b8
Changed mpl backend to TkAgg
