### 2019-11-13 d197f25
Update matplotlib import for mapping
