+++ .github/workflows/pythonpackage.yml
@@ -30,8 +30,9 @@ jobs:
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test (WIP)
-      run: python -m unittest discover -p "utest_ROI_analysis*.py", "utest_SAR_calc*.py" 
- 
+      run: python -m unittest discover -p "utest_ROI_analysis*.py"
+    - name: Test (WIP2)
+      run: python -m unittest discover -p  "utest_SAR_calc*.py" 
       #run: python -m unittest
       #pip install pytest
       #pytest
