### 2019-11-13 c0c2e36
Install PyPulseq from latest Github commit
