### 2019-11-13 5d5e300
Install PyPulseq 1.0.0r1
PyPulseq 1.0.0r1 fixes Matplotlib import issues on Virtual Scanner
