+++ requirements.txt
@@ -24,7 +24,7 @@ Pillow==6.0.0
 protobuf==3.7.1
 pydicom==1.2.2
 pyparsing==2.3.1
-git+https://github.com/imr-framework/pypulseq.git
+pypulseq==1.0.0post1
 python-dateutil==2.8.0
 PyYAML==5.1
 scipy==1.2.1
