### 2019-11-13 0ec6e47
Bump pillow from 6.0.0 to 6.2.0
Bumps [pillow](https://github.com/python-pillow/Pillow) from 6.0.0 to 6.2.0.
- [Release notes](https://github.com/python-pillow/Pillow/releases)
- [Changelog](https://github.com/python-pillow/Pillow/blob/master/CHANGES.rst)
- [Commits](https://github.com/python-pillow/Pillow/compare/6.0.0...6.2.0)

Signed-off-by: dependabot[bot] <support@github.com>
