+++ requirements.txt
@@ -20,7 +20,7 @@ nibabel==2.4.0
 numpy==1.16.2
 opencv-python==4.0.0.21
 pbr==5.1.3
-Pillow==6.0.0
+Pillow==6.2.0
 protobuf==3.7.1
 pydicom==1.2.2
 pyparsing==2.3.1
