### 2019-11-13 6537cf2
Modify mapping code
