+++ .github/workflows/pythonpackage.yml
@@ -1,6 +1,9 @@
 name: Python package
 
-on: [push]
+on: 
+  push:
+    branches: 
+      - master
 
 jobs:
   build:
