+++ .github/workflows/pythonpackage_dev.yml
@@ -1,0 +1,40 @@
+name: Python package
+
+on: 
+  push:
+    branches:
+      - dev
+
+jobs:
+  build:
+   # runs-on: windows-latest
+    runs-on: ${{ matrix.os }}
+    strategy:
+      max-parallel: 2
+      matrix:
+        #os: [windows-latest]
+        os: [macOS-latest, windows-latest]
+   
+    steps:
+    - uses: actions/checkout@v1
+    - name: Set up Python 3.6
+      uses: actions/setup-python@v1
+      with:
+        python-version: 3.6
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
+    - name: Lint with flake8
+      run: |
+        pip install flake8
+        # stop the build if there are Python syntax errors or undefined names
+        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+        flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+    - name: Test (WIP)
+      run: python -m unittest discover -p "utest*.py"
+
+      #run: python -m unittest
+      #pip install pytest
+      #pytest
