+++ .github/workflows/pythonpackage_dev.yml
@@ -32,8 +32,8 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-    - name: Test (WIP)
-      run: python -m unittest discover -p "utest*.py"
+    - name: Test (T1/T2 mapping only)
+      run: python -m unittest discover -p "utest_T1_T2_mapping*.py"
 
       #run: python -m unittest
       #pip install pytest
