### 2019-11-13 6dece64
Change imread to imageio.imread
