+++ virtualscanner/server/ana/utest_T1_T2_mapping.py
@@ -7,6 +7,7 @@ import virtualscanner.server.ana.T2_mapping as dicom2mapT2
 import numpy as np
 from scipy.misc import imread
 from virtualscanner.utils import constants
+import imageio
 
 SERVER_T1_INPUT_PATH = constants.SERVER_ANALYZE_PATH / 'inputs' / 'T1_orig_data'
 SERVER_T2_INPUT_PATH = constants.SERVER_ANALYZE_PATH / 'inputs' / 'T2_orig_data'
@@ -32,8 +33,8 @@ class MyTestCase(unittest.TestCase):
                     'linux': SERVER_T1_MAC_MAP_PATH}
         dicom_map_path = platform[sys.platform]
         map_name, dicom_path = dicom2mapT1.main(dicom_file_path=SERVER_T1_INPUT_PATH , TR=TRstr, TE=TEstr, TI=TIstr, pat_id='9306')
-        generated_map = imread(COMS_MAP_PATH / map_name)
-        utest_map = imread(dicom_map_path / 'utest_T1_map.png')
+        generated_map = imageio.imread(COMS_MAP_PATH / map_name)
+        utest_map = imageio.imread(dicom_map_path / 'utest_T1_map.png')
 
         np.testing.assert_allclose(generated_map, utest_map)
 
@@ -46,8 +47,8 @@ class MyTestCase(unittest.TestCase):
                     'linux': SERVER_T2_MAC_MAP_PATH}
         dicom_map_path = platform[sys.platform]
         map_name, dicom_path = dicom2mapT2.main(dicom_file_path=SERVER_T2_INPUT_PATH, TR=TRstr, TE=TEstr, pat_id='9306')
-        generated_map = imread(COMS_MAP_PATH / map_name)
-        utest_map = imread(dicom_map_path / 'utest_T2_map.png')
+        generated_map = imageio.imread(COMS_MAP_PATH / map_name)
+        utest_map = imageio.imread(dicom_map_path / 'utest_T2_map.png')
 
         np.testing.assert_allclose(generated_map, utest_map)
 
