### 2019-11-14 751481e
Attempt to debug mapping code
