+++ virtualscanner/server/ana/utest_T1_T2_mapping.py
@@ -35,7 +35,8 @@ class MyTestCase(unittest.TestCase):
         map_name, dicom_path = dicom2mapT1.main(dicom_file_path=SERVER_T1_INPUT_PATH , TR=TRstr, TE=TEstr, TI=TIstr, pat_id='9306')
         generated_map = imageio.imread(COMS_MAP_PATH / map_name)
         utest_map = imageio.imread(dicom_map_path / 'utest_T1_map.png')
-
+        utest_map = utest_map[0:487, :, :]
+        
         np.testing.assert_allclose(generated_map, utest_map)
 
     def test_T2_mapping(self):
@@ -49,6 +50,7 @@ class MyTestCase(unittest.TestCase):
         map_name, dicom_path = dicom2mapT2.main(dicom_file_path=SERVER_T2_INPUT_PATH, TR=TRstr, TE=TEstr, pat_id='9306')
         generated_map = imageio.imread(COMS_MAP_PATH / map_name)
         utest_map = imageio.imread(dicom_map_path / 'utest_T2_map.png')
+        utest_map = utest_map[0:487, :, :]
 
         np.testing.assert_allclose(generated_map, utest_map)
 
