### 2019-11-14 7b9e64c
Attempt to debug mapping code
