### 2019-11-14 2a8b40d
Attempt to debug mapping code
