+++ .github/workflows/pythonpackage_dev.yml
@@ -12,8 +12,8 @@ jobs:
     strategy:
       max-parallel: 2
       matrix:
-        #os: [windows-latest]
-        os: [macOS-latest, windows-latest]
+        os: [windows-latest]
+#        os: [macOS-latest, windows-latest]
    
     steps:
     - uses: actions/checkout@v1
