+++ .github/workflows/pythonpackage.yml
@@ -33,10 +33,10 @@ jobs:
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test (WIP)
-      run: python "coms_server_flask.py"
+      run: python "./virtualscanner/coms/coms_ui/coms_server_flask.py"
         #run: python -m unittest discover -p "utest*.py"
     - name: Test (WIPUI)
-      run: python "standard_mode_GUI_selenium.py"
+      run: python "./virtualscanner/coms/coms_ui/standard_mode_GUI_selenium.py"
       #run: python -m unittest
       #pip install pytest
       #pytest
