+++ virtualscanner/server/ana/utest_T1_T2_mapping.py
@@ -5,7 +5,6 @@ import unittest
 import virtualscanner.server.ana.T1_mapping as dicom2mapT1
 import virtualscanner.server.ana.T2_mapping as dicom2mapT2
 import numpy as np
-from matplotlib.pyplot import imread
 from virtualscanner.utils import constants
 import imageio
 
@@ -15,6 +14,8 @@ SERVER_T1_WINDOWS_MAP_PATH = constants.SERVER_ANALYZE_PATH / 'inputs' / 'T1_map_
 SERVER_T1_MAC_MAP_PATH = constants.SERVER_ANALYZE_PATH / 'inputs' / 'T1_map_mac'
 SERVER_T2_WINDOWS_MAP_PATH = constants.SERVER_ANALYZE_PATH / 'inputs' / 'T2_map_windows'
 SERVER_T2_MAC_MAP_PATH = constants.SERVER_ANALYZE_PATH / 'inputs' / 'T2_map_mac'
+SERVER_T1_MAP_PATH = constants.SERVER_ANALYZE_PATH / 'outputs' / '9306' / 'T1_map'
+SERVER_T2_MAP_PATH = constants.SERVER_ANALYZE_PATH / 'outputs' / '9306' / 'T2_map'
 COMS_MAP_PATH = constants.COMS_UI_PATH / 'static' / 'ana' / 'outputs' / '9306'
 
 
@@ -32,9 +33,9 @@ class MyTestCase(unittest.TestCase):
                     'darwin': SERVER_T1_MAC_MAP_PATH,
                     'linux': SERVER_T1_MAC_MAP_PATH}
         dicom_map_path = platform[sys.platform]
-        map_name, dicom_path = dicom2mapT1.main(dicom_file_path=SERVER_T1_INPUT_PATH , TR=TRstr, TE=TEstr, TI=TIstr, pat_id='9306')
-        generated_map = imageio.imread(str(COMS_MAP_PATH / map_name))
-        utest_map = imageio.imread(str(dicom_map_path / 'utest_T1_map.png'))
+        map_name, dicom_path, np_map_name = dicom2mapT1.main(dicom_file_path=SERVER_T1_INPUT_PATH , TR=TRstr, TE=TEstr, TI=TIstr, pat_id='9306')
+        generated_map = np.load(SERVER_T1_MAP_PATH / np_map_name)
+        utest_map = np.load(dicom_map_path / 'utest_T1_map.npy')
 
         np.testing.assert_allclose(generated_map, utest_map)
 
@@ -46,10 +47,9 @@ class MyTestCase(unittest.TestCase):
                     'darwin': SERVER_T2_MAC_MAP_PATH,
                     'linux': SERVER_T2_MAC_MAP_PATH}
         dicom_map_path = platform[sys.platform]
-        print(dicom_map_path)
-        map_name, dicom_path = dicom2mapT2.main(dicom_file_path=SERVER_T2_INPUT_PATH, TR=TRstr, TE=TEstr, pat_id='9306')
-        generated_map = imageio.imread(str(COMS_MAP_PATH / map_name))
-        utest_map = imageio.imread(str(dicom_map_path / 'utest_T2_map.png'))
+        map_name, dicom_path, np_map_name = dicom2mapT2.main(dicom_file_path=SERVER_T2_INPUT_PATH, TR=TRstr, TE=TEstr, pat_id='9306')
+        generated_map = np.load(SERVER_T2_MAP_PATH / np_map_name)
+        utest_map = np.load(dicom_map_path / 'utest_T2_map.npy')
 
         np.testing.assert_allclose(generated_map, utest_map)
 
