+++ .github/workflows/pythonpackage_dev.yml
@@ -13,8 +13,8 @@ jobs:
       max-parallel: 2
       matrix:
 #        os: [windows-latest]
-         os: [macOS-Latest]
-#        os: [macOS-latest, windows-latest]
+#         os: [macOS-Latest]
+        os: [macOS-latest, windows-latest]
    
     steps:
     - uses: actions/checkout@v1
