### 2019-11-16 9e81471
Add GUI test functions (not in unittest format)
