### 2019-11-18 #1637 @tonggehua
@labarba We have got the front end testing working on Windows.  Currently, we are wrapping it up in a test case for Github actions to run it along with all the back end tests. The goal is to have this done by Friday.
