+++ virtualscanner/server/ana/utest_T1_T2_mapping.py
@@ -39,7 +39,8 @@ class MyTestCase(unittest.TestCase):
 
         map_diff = np.subtract(generated_map, utest_map)
         non_zero_loc = np.nonzero(map_diff)
-        print(non_zero_loc.shape)
+        non_zero_loc_np = np.asarray(non_zero_loc)
+        print(non_zero_loc_np.shape)
         print(non_zero_loc)
 
         np.testing.assert_allclose(generated_map, utest_map)
@@ -58,7 +59,8 @@ class MyTestCase(unittest.TestCase):
 
         map_diff = np.subtract(generated_map, utest_map)
         non_zero_loc = np.nonzero(map_diff)
-        print(non_zero_loc.shape)
+        non_zero_loc_np = np.asarray(non_zero_loc)
+        print(non_zero_loc_np.shape)
         print(non_zero_loc)
 
         np.testing.assert_allclose(generated_map, utest_map)
