+++ .github/workflows/pythonpackage_dev.yml
@@ -7,7 +7,6 @@ on:
 
 jobs:
   build:
-   # runs-on: windows-latest
     runs-on: ${{ matrix.os }}
     strategy:
       max-parallel: 2
@@ -18,6 +17,8 @@ jobs:
    
     steps:
     - uses: actions/checkout@v1
+    - name: Setup tmate session
+      uses: mxschmitt/action-tmate@v1
     - name: Set up Python 3.6
       uses: actions/setup-python@v1
       with:
@@ -36,6 +37,3 @@ jobs:
     - name: Test (T1/T2 mapping only)
       run: python -m unittest discover -p "utest_T1_T2_mapping*.py"
 
-      #run: python -m unittest
-      #pip install pytest
-      #pytest
