+++ .github/workflows/pythonpackage_dev.yml
@@ -17,8 +17,6 @@ jobs:
    
     steps:
     - uses: actions/checkout@v1
-    - name: Setup tmate session
-      uses: mxschmitt/action-tmate@v1
     - name: Set up Python 3.6
       uses: actions/setup-python@v1
       with:
@@ -34,6 +32,8 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+    - name: Setup tmate session
+      uses: mxschmitt/action-tmate@v1
     - name: Test (T1/T2 mapping only)
       run: python -m unittest discover -p "utest_T1_T2_mapping*.py"
 
