### 2019-11-18 9de96e3
Test with selenium (works on Windows)
These two scripts work together to do a front end test; GUI_test_functions.py contains the actual test class and tests,  while utest_GUI.py opens a side thread that runs the tests while the main thread runs the app from coms_server_flask.py; all 7 tests have passed on Windows 10 pro (local)
