+++ virtualscanner/coms/coms_ui/utest_GUI.py
@@ -7,6 +7,10 @@ import threading
 from virtualscanner.coms.coms_ui import GUI_test_functions
 from virtualscanner.coms.coms_ui import coms_server_flask
 import time
+import unittest
+from pprint import pprint
+
+
 # import webbrowser
 # import subprocess
 # from virtualscanner.utils import constants
@@ -16,10 +20,50 @@ def selenium_function():
     time.sleep(10)
     # This URL needs to be changed to the server address if running on a remote client on the local network
     # webbrowser.open("http://0.0.0.0:5000/")
-    GUI_test_functions.launch_tests()
+   ###
+   # return GUI_test_functions.launch_tests()
+   ###
+    runner = unittest.TextTestRunner()
+    result = runner.run(unittest.makeSuite(GUI_test_functions.GUItestclass))
+
     # Do all tests as you want here and get 200 responses.
     # Figure out how to report 200 responses if required
     # webbrowser.close()
+    return result
+
+
+#class GUI_test_case(unittest.TestCase):
+ #   t = threading.Thread(target=selenium_function)
+class ThreadWithReturn(threading.Thread):
+    # Copied from stack overflow question # 6893968
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._return = None
+
+    def run(self):
+        target = getattr(self, '_target')
+        if not target is None:
+            self._return = target(*getattr(self, '_args'), **getattr(self,'_kwargs'))
+
+    def join(self, *args, **kwargs):
+        super().join(*args, **kwargs)
+        return self._return
+#
+#
+
+# class TestGUIFunctions(unittest.TestCase):
+#     def test_gui_all(self):
+#         t_sel = ThreadWithReturn(target=selenium_function)
+#         t_sel.daemon = True
+#
+#
+#         t_sel.start()
+#         coms_server_flask.launch_virtualscanner()
+#
+# #        coms_server_flask.launch_virtualscanner()
+#         self.assertEqual(t_sel.join(), "Test Code")
+
+
 
 
 if __name__ == '__main__':
@@ -27,3 +71,5 @@ if __name__ == '__main__':
     t.daemon = True
     t.start()
     coms_server_flask.launch_virtualscanner()
+   #unittest.main()
+
