+++ virtualscanner/coms/coms_ui/GUI_test_functions.py
@@ -12,208 +12,264 @@ import unittest
 from selenium.webdriver.common.action_chains import ActionChains
 import webbrowser
 
-
+from virtualscanner.utils.constants import COMS_UI_PATH
 
 #subprocess.call(['python', str(constants.COMS_UI_PATH / 'coms_server_flask.py')], shell=True)
 
 # Create a new instance of the Firefox driver
 
 
-def get_vs_address():
-    print(sys.platform)
-    if sys.platform == 'win32':
-        vs_address = "http://127.0.0.1:5000/"
-    else:
-        vs_address = "http://0.0.0.0:5000/"
+class GUItestclass(unittest.TestCase):
 
-    return vs_address
+    def get_vs_address(self):
+        print(sys.platform)
+        if sys.platform == 'win32':
+            vs_address = "http://127.0.0.1:5000/"
+        else:
+            vs_address = "http://0.0.0.0:5000/"
 
-def setup():
-    driver = webdriver.Firefox(executable_path="./firefox_driver/geckodriver.exe")
- # Go to Virtual Scanner
-    driver.get(get_vs_address())
-    return driver
+        return vs_address
 
-def run_test(test, expected_value):
+    def setup_driver(self):
+        if sys.platform == 'win32':
+            driver_path = str(COMS_UI_PATH / "firefox_driver/geckodriver_windows.exe")
+        else:
+            driver_path = str(COMS_UI_PATH / "firefox_driver/geckodriver_mac")
+        print("Firefox driver path: ", driver_path)
+        driver = webdriver.Firefox(executable_path=driver_path)
+        # Go to Virtual Scanner
 
-    driver = setup()
+        return driver
 
-    value = test(driver)
-    print("Expected value: ", expected_value)
-    print("Test value: ", value)
+    def run_test(self, test, expected_value):
+        driver = self.setUp()
+        value = test(driver)
+        print("Expected value: ", expected_value)
+        print("Test value: ", value)
+        assert value == expected_value, "Value is not the same as expected :("
+        driver.quit()
 
-    assert value == expected_value, "Value is not the same as expected :("
-    print("")
-    driver.quit()
-
-
-def login_std(driver):
-    # What does an implicit wait do?
-    driver.find_element_by_id("mode-selection").submit()
-    time.sleep(3)
-
-    return driver.current_url
+        return value==expected_value
 
 
-def login_adv(driver):
-    driver.find_element_by_id("opt2").click()
-    time.sleep(1)
-    driver.find_element_by_id("mode-selection").submit()
-    time.sleep(3)
-
-    return driver.current_url
-
-
-def register(driver):
-    # Login first
-    login_std(driver)
-
-    # Find register button and click
-    driver.find_element_by_id("reg-form-submit").click()
-    time.sleep(3)
-
-    # Check and return style of element with id="success-sentence"
-    success_sentence = driver.find_element_by_id("success-sentence")
-    time.sleep(3)
-
-    return success_sentence.get_attribute("style")
-
-
-def acquire(driver):
-    register(driver)
-
-    driver.find_element_by_link_text("Acquire").click()
-    time.sleep(2)
-
-    orientations = ['axial']
-    seq_types = ['GRE']
-
-    for orientation in orientations:
-        for seq_type in seq_types:
-            ir = False
-            if seq_type == 'IRSE':
-                seq_name = 'SE'
-                ir = True
-            else:
-                seq_name = seq_type
-
-            # Add sequence
-            driver.find_element_by_id("addseq-btn").click() #
-            time.sleep(1)
-            driver.find_element_by_link_text(seq_name).click()
-            time.sleep(1)
-
-            # Change parameters for faster sim (5 x 5)
-            inputnx = driver.find_element_by_id("Nx")
-            inputnx.clear()
[…]
