+++ .github/workflows/pythonpackage.yml
@@ -32,11 +32,9 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-    - name: Test (WIP)
-      run: python "./virtualscanner/coms/coms_ui/coms_server_flask.py"
-        #run: python -m unittest discover -p "utest*.py"
-    - name: Test (WIPUI)
-      run: python "./virtualscanner/coms/coms_ui/standard_mode_GUI_selenium.py"
+    - name: Test (WIP - GUI only)
+      #run: python "./virtualscanner/coms/coms_ui/coms_server_flask.py"
+      run: python -m unittest discover -p "utest_GUI*.py"
       #run: python -m unittest
       #pip install pytest
       #pytest
