### 2019-11-18 eb2e4c9
Update requirements to include Selenium
