+++ .github/workflows/pythonpackage.yml
@@ -33,8 +33,9 @@ jobs:
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test (WIP - GUI only)
+      run: python "./virtualscanner/coms/coms_ui/utest_GUI.py"
       #run: python "./virtualscanner/coms/coms_ui/coms_server_flask.py"
-      run: python -m unittest discover -p "utest_GUI*.py"
+      #run: python -m unittest discover -p "utest_GUI*.py"
       #run: python -m unittest
       #pip install pytest
       #pytest
