+++ .github/workflows/pythonpackage.yml
@@ -32,11 +32,11 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-    - name: Test (WIP - coms server flask)
-      #run: python "./virtualscanner/coms/coms_ui/utest_GUI.py"
-      run: python "./virtualscanner/coms/coms_ui/coms_server_flask.py"
-    - name: Test (WIP - GUI with unittest?)
-      run: python -m unittest discover -p "utest_GUI.py"
+    - name: Test (WIP - GUI test)
+      run: python "./virtualscanner/coms/coms_ui/GUI_utest.py"
+      #run: python "./virtualscanner/coms/coms_ui/coms_server_flask.py"
+  #  - name: Test (WIP - GUI with unittest?)
+      #run: python -m unittest discover -p "utest_GUI.py"
       #run: python -m unittest
       #pip install pytest
       #pytest
