### 2019-11-19 16df648
MacOS run with GUI testing
