+++ .github/workflows/pythonpackage.yml
@@ -12,9 +12,8 @@ jobs:
     strategy:
       max-parallel: 2
       matrix:
-        os: [windows-latest]
-        #os: [macOS-latest, windows-latest]
-   
+        os: [macOS-latest]      
+        #os: [windows-latest]
     steps:
     - uses: actions/checkout@v1
     - name: Set up Python 3.6
