+++ virtualscanner/coms/coms_ui/GUI_utest.py
@@ -4,8 +4,10 @@ This script unit starts and tests the communications between server and client(s
 """
 
 import threading
-from virtualscanner.coms.coms_ui.GUI_test_functions import GUItestclass
-from virtualscanner.coms.coms_ui.coms_server_flask import launch_virtualscanner
+#from virtualscanner.coms.coms_ui.GUI_test_functions import GUItestclass
+from GUI_test_functions import GUItestclass
+#from virtualscanner.coms.coms_ui.coms_server_flask import launch_virtualscanner
+from coms_server_flask import launch_virtualscanner
 import time
 import unittest
 from pprint import pprint
