### 2019-11-20 283352f
Check if multithread is working on Github actions
