+++ .github/workflows/pythonpackage.yml
@@ -33,7 +33,7 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Enable macOS Firefox driver
       if: matrix.os == 'macOS-latest'
-      run: sudo chmod +x "./virtualscanner/coms/coms_ui/geckodriver_mac"
+      run: sudo chmod +x "./virtualscanner/coms/coms_ui/firefox_driver/geckodriver_mac"
     - name: Test (WIP - GUI test)
       run: python "./virtualscanner/coms/coms_ui/GUI_utest.py"
   #  - name: Test (WIP - GUI with unittest?)
