### 2019-11-20 302456b
Add a definition to terminate the server
