### 2019-11-20 imr-framework/virtual-scanner#38 @tonggehua
**Differences between local testing and Github Actions testing**

@vsoch We are tying up the Selenium testing. The way we are doing it right now is to run `coms_server_flask.py` while another thread kicks off a unittest runner that uses Selenium to click elements. At the moment, the tests check whether you are on the right page / correct elements are present, for login/register/acquire/analzye/tx/rx.

We are seeing discrepancies between local runs and Github action runs:

1. While the GUI test runs successfully on local, Github Actions Selenium has problems clicking elements (on Acquire and Analyze tests that run the computations) because some other element is obscuring them. The problem is probably window size, and it makes less sense to debug it now that a new team has committed to developing a more mature GUI with us.

2. T1/T2 mapping tests compare generated data to benchmark data. The tests pass on Windows (both local and Github actions) and macOS (local), but generates a much larger error on macOS (Github actions).  This is described in issue #41.

3. Github Actions macOS doesn't have Firefox installed for Selenium testing.

Would this be acceptable right now?
