### 2019-11-20 imr-framework/virtual-scanner#38 @vsoch
The one thing I would try before completely throwing in the towel is to run the same tests, but use CircleCI. I've done quite a bit of browser testing there and consistently get the same tests locally as on the CI. After that, if there is still issue, then we can say that your team gave their best effort and no go!

But maybe we will get lucky and the CircleCI runs will be consistent!
