+++ .github/workflows/pythonpackage_dev.yml
@@ -11,8 +11,6 @@ jobs:
     strategy:
       max-parallel: 2
       matrix:
-#        os: [macOS-latest]
-#        os: [windows-latest]
         os: [macOS-latest, windows-latest]
    
     steps:
@@ -34,10 +32,5 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test (T1/T2 mapping only)
       run: python -m unittest discover -p "utest_T1_T2_mapping*.py"
-    - uses: actions/download-artifact@master
-      with:
-        name: my-artifact
-        path: path/to/artifact
 
-    - run: cat path/to/artifact
 
