### 2019-11-20 imr-framework/virtual-scanner#38 @sairamgeethanath
@vsoch We have trying to get this done on Github Actions and want to mark the rest as "To Dos" for the future. It will be another effort to move this to CircleCI. The tests run locally on both operating systems. I appreciate your recommendations but moving to another tool all over will lose all the effort put on Github actions and require time and effort again. I kindly request you to reconsider your suggestion.
