### 2019-11-20 imr-framework/virtual-scanner#38 @vsoch
okay, sure thing then. The tests aren't automated and this is below my personal standard (and review criteria) but I can make an exception because I know the amount of time and work your team has put into this. I won't check the automated tests box or close this issue, but I'll let the editors know that I'm consenting to this.
