### 2019-11-20 #1637 @vsoch
@labarba the authors have tests that work locally but not on CI (not automated), and have requested that the requirement is waived. This is below my personal standard and review criteria, but I do recognize they put some effort into trying. I won't check the automated tests box or close this issue, but I'll forward the final decision on to you.
