### 2019-11-20 imr-framework/virtual-scanner#38 @sairamgeethanath
Thanks for your comments! However, I would like to clarify that the Github actions are successfully running automated tests for backend functionalities. The front end GUI tests based on your Selenium suggestion runs for windows as well. We are pointing to the discrepancies between local and Github Actions (something we cannot control much). We are not sure about moving to other CI platforms as Actions/related features are well integrated in one tool that we are using. Many thanks for your consideration!
