### 2019-11-20 imr-framework/virtual-scanner#38 @vsoch
> The front end GUI tests based on your Selenium suggestion runs for windows as well.

Yes, and hopefully this can be fully finished at some point to be run automated.

I suggested in the issue to use Selenium on Circle, which has worked for me many times, and the authors asked for a waiver.
