### 2019-11-20 #1637 @sairamgeethanath
@labarba Pl. also refer our response: https://github.com/imr-framework/virtual-scanner/issues/38#issuecomment-556128982
