### 2019-11-20 imr-framework/virtual-scanner#38 @sairamgeethanath
Thanks, just to clarify:
This means looking at the discrepancies between Github Actions (different python versions, configurations) and local testing. I am not sure if we need to focus on the software version/libraries discrepancies. Circle CI does not offer "Mac" for the free version.
