### 2019-11-20 #1637 @arokem
I think that based on our [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) this falls in the "OK" category.
