+++ virtualscanner/server/ana/utest_T1_T2_mapping.py
@@ -37,11 +37,6 @@ class MyTestCase(unittest.TestCase):
         generated_map = np.load(SERVER_T1_MAP_PATH / np_map_name)
         utest_map = np.load(dicom_map_path / 'utest_T1_map.npy')
 
-        # map_diff = np.subtract(generated_map, utest_map)
-        # non_zero_loc = np.nonzero(map_diff)
-        # non_zero_loc_np = np.asarray(non_zero_loc)
-        # print(non_zero_loc_np.shape)
-        # print(non_zero_loc)
         python_version = sys.version_info
         if python_version.major == 3 and python_version.minor == 6 and python_version.micro == 9:
             rtol = 30
@@ -63,11 +58,6 @@ class MyTestCase(unittest.TestCase):
         generated_map = np.load(SERVER_T2_MAP_PATH / np_map_name)
         utest_map = np.load(dicom_map_path / 'utest_T2_map.npy')
 
-        # map_diff = np.subtract(generated_map, utest_map)
-        # non_zero_loc = np.nonzero(map_diff)
-        # non_zero_loc_np = np.asarray(non_zero_loc)
-        # print(non_zero_loc_np.shape)
-        # print(non_zero_loc)
         python_version = sys.version_info
         if python_version.major == 3 and python_version.minor == 6 and python_version.micro == 9:
             rtol = 30
@@ -76,7 +66,6 @@ class MyTestCase(unittest.TestCase):
             rtol = 1e-7
             atol = 0
         np.testing.assert_allclose(generated_map, utest_map, rtol, atol)
-
 
 if __name__ == '__main__':
     unittest.main()
