### 2019-11-20 2771c92
Add def to terminate utest
