### 2019-11-20 32bf3e2
Only preserve back-end automatic tests
