### 2019-11-21 #1637 @kthyng
Another note: Your paper looks good, but the capitalization in your references is off. When you want to preserve capitalization in bibtex references, you can add {} around the word or phrase. For example, I see "nmr" several times and I believe this should be capitalized. You can preserve that with {NMR} in the bibtex entry.  Could you go through your references and make sure they come through to the paper looking correct in the references list?
