### 2019-11-21 0d8769e
Fix upper-case initialisms in references
