+++ paper/paper.bib
@@ -13,7 +13,7 @@
 
 
 @article{ravi2018pulseq,
- title={Pulseq-Graphical Programming Interface: Open source visual environment for prototyping pulse sequences and integrated magnetic resonance imaging algorithm development},
+ title={{P}ulseq-{G}raphical {P}rogramming {I}nterface: Open source visual environment for prototyping pulse sequences and integrated magnetic resonance imaging algorithm development},
  author={Ravi, Keerthi Sravan and Potdar, Sneha and Poojar, Pavan and Reddy, Ashok Kumar and Kroboth, Stefan and Nielsen, Jon-Fredrik and Zaitsev, Maxim and Venkatesan, Ramesh and Geethanath, Sairam},
  journal={Magnetic resonance imaging},
  volume={52},
@@ -27,7 +27,7 @@
 @misc{obungoloch_2019, title={2019 Junior Fellow Challenge: Africa}, url={https://www.ismrm.org/2019-junior-fellow-challenge/africa/}, journal={ISMRM}, author={Obungoloch, Johnes}, year={2019}}
 
 @article{kose2017blochsolver,
-  title={BlochSolver: a GPU-optimized fast 3D MRI simulator for experimentally compatible pulse sequences},
+  title={BlochSolver: a GPU-optimized fast 3D {MRI} simulator for experimentally compatible pulse sequences},
   author={Kose, Ryoichi and Kose, Katsumi},
   journal={Journal of Magnetic Resonance},
   volume={281},
@@ -43,7 +43,7 @@
 @article{Graesslin_SAR,
 doi={10.1002/mrm.24138},
 author = {Graesslin, Ingmar and Homann, Hanno and Biederer, Sven and Börnert, Peter and Nehrke, Kay and Vernickel, Peter and Mens, Giel and Harvey, Paul and Katscher, Ulrich},
-title = {A specific absorption rate prediction concept for parallel transmission M{R}},
+title = {A specific absorption rate prediction concept for parallel transmission {MR}},
 journal = {Magnetic Resonance in Medicine},
 volume = {68},
 number = {5},
@@ -57,12 +57,12 @@ year = {2012}
 
 @book{webb2016magnetic,
   
-title={Magnetic Resonance Technology: Hardware and System Component Design},
+title={Magnetic {R}esonance {T}echnology: Hardware and {S}ystem {C}omponent {D}esign},
   
 author={Webb, Andrew G.},
   isbn={9781782623595},
   lccn={2016439714},
-  series={New Developments in NMR},
+  series={New {D}evelopments in {NMR}},
   url={https://books.google.com/books?id=PYkmDAAAQBAJ},
   year={2016},
   publisher={Royal Society of Chemistry}
@@ -70,7 +70,7 @@ author={Webb, Andrew G.},
 
 
 @article{mugler1990three,
-  title={Three-dimensional magnetization-prepared rapid gradient-echo imaging (3D M{P} R{AGE})},
+  title={Three-dimensional magnetization-prepared rapid gradient-echo imaging (3D {MP} {RAGE})},
   author={Mugler III, John P and Brookeman, James R},
   journal={Magnetic Resonance in Medicine},
   volume={15},
@@ -106,7 +106,7 @@ doi={10.1126/science.1925560}
 }
 
 @article{stocker2010high,
-  title={High-performance computing M{RI} simulations},
+  title={High-performance computing {MRI} simulations},
   author={St{\"o}cker, Tony and Vahedipour, Kaveh and Pflugfelder, Daniel and Shah, N Jon},
   journal={Magnetic resonance in medicine},
   volume={64},
@@ -120,7 +120,7 @@ doi={10.1126/science.1925560}
[…]
