### 2019-11-21 b6d39ab
Fix capitalizations (II)
