+++ paper/paper.bib
@@ -2,7 +2,7 @@
 @article{layton2017pulseq,
   title={Pulseq: a rapid and hardware-independent pulse sequence prototyping framework},
   author={Layton, Kelvin J and Kroboth, Stefan and Jia, Feng and Littin, Sebastian and Yu, Huijun and Leupold, Jochen and Nielsen, Jon-Fredrik and St{\"o}cker, Tony and Zaitsev, Maxim},
-  journal={Magnetic resonance in medicine},
+  journal={Magnetic Resonance in Medicine},
   volume={77},
   number={4},
   pages={1544--1552},
@@ -15,7 +15,7 @@
 @article{ravi2018pulseq,
  title={{P}ulseq-{G}raphical {P}rogramming {I}nterface: Open source visual environment for prototyping pulse sequences and integrated magnetic resonance imaging algorithm development},
  author={Ravi, Keerthi Sravan and Potdar, Sneha and Poojar, Pavan and Reddy, Ashok Kumar and Kroboth, Stefan and Nielsen, Jon-Fredrik and Zaitsev, Maxim and Venkatesan, Ramesh and Geethanath, Sairam},
- journal={Magnetic resonance imaging},
+ journal={Magnetic Resonance Imaging},
  volume={52},
  pages={9--15},
  year={2018},
@@ -27,7 +27,7 @@
 @misc{obungoloch_2019, title={2019 Junior Fellow Challenge: Africa}, url={https://www.ismrm.org/2019-junior-fellow-challenge/africa/}, journal={ISMRM}, author={Obungoloch, Johnes}, year={2019}}
 
 @article{kose2017blochsolver,
-  title={BlochSolver: a GPU-optimized fast 3D {MRI} simulator for experimentally compatible pulse sequences},
+  title={BlochSolver: a {GPU}-optimized fast 3{D} {MRI} simulator for experimentally compatible pulse sequences},
   author={Kose, Ryoichi and Kose, Katsumi},
   journal={Journal of Magnetic Resonance},
   volume={281},
@@ -44,7 +44,7 @@
 doi={10.1002/mrm.24138},
 author = {Graesslin, Ingmar and Homann, Hanno and Biederer, Sven and Börnert, Peter and Nehrke, Kay and Vernickel, Peter and Mens, Giel and Harvey, Paul and Katscher, Ulrich},
 title = {A specific absorption rate prediction concept for parallel transmission {MR}},
-journal = {Magnetic Resonance in Medicine},
+journal = {{M}agnetic {R}esonance in {M}edicine},
 volume = {68},
 number = {5},
 pages = {1664-1674},
@@ -108,7 +108,7 @@ doi={10.1126/science.1925560}
 @article{stocker2010high,
   title={High-performance computing {MRI} simulations},
   author={St{\"o}cker, Tony and Vahedipour, Kaveh and Pflugfelder, Daniel and Shah, N Jon},
-  journal={Magnetic resonance in medicine},
+  journal={Magnetic {R}esonance in {M}edicine},
   volume={64},
   number={1},
   pages={186--193},
@@ -122,7 +122,7 @@ doi={10.1126/science.1925560}
 @article{zwart2015graphical,
   title={Graphical programming interface: A development environment for {MRI} methods},
   author={Zwart, Nicholas R and Pipe, James G},
-  journal={Magnetic resonance in medicine},
+  journal={{M}agnetic {R}esonance in {M}edicine},
   volume={74},
   number={5},
   pages={1449--1460},
@@ -141,7 +141,7 @@ doi={10.1126/science.1925560}
 }
 
 @book{brown2014magnetic,
[…]
