+++ paper/paper.bib
@@ -194,7 +194,8 @@ url={https://doi.org/10.1007/s12045-015-0268-2}
   author={McRobbie, Donald W and Moore, Elizabeth A and Graves, Martin J and Prince, Martin R},
   year={2017},
   publisher={Cambridge {U}niversity {P}ress},
-  chapter={2}
+  chapter={2},
+  doi={10.1016/j.crad.2003.09.015}
 }
 
 @book{brown2015mri,
