### 2019-11-25 #1637 @kthyng
@imr-framework, @tonggehua I have added the doi stated as missing in the previous references check in PR [43](https://github.com/imr-framework/virtual-scanner/pull/43) if you want to merge
