### 2019-11-25 #1637 @tonggehua
@kthyng Thank you! I merged it.
